API Reference

Ray Interface

Base.eltype — Method

Base.eltype(::Type{<:AbstractRay})
Base.eltype(::AbstractRay)

Return the numeric type used by a ray object or a ray type.

Base.ndims — Method

Base.ndims(::Type{<:AbstractRay})
Base.ndims(::AbstractRay)

Return the number of dimensions of a ray object or a ray type.

RayTracing.advance! — Method

advance!(ray, distance)

Advances the ray by the given distance in its current direction.

RayTracing.direction — Function

direction(ray)

Return the current direction of the ray, as a mutable reference.

RayTracing.position — Function

Base.position(ray)
RayTracing.position(ray)

Return the current position of the ray, as a mutable reference.

RayTracing.data! — Method

data!(ray::RayWithMutableData, data)

Set the data carried by ray.

RayTracing.data — Method

data(ray::AbstractRayWithData)

Return the data carried by ray.

Ray Types

RayTracing.AbstractRay — Type

abstract type AbstractRay{ND, T <: Real}

Represents a ray in ND dimensions that uses the type T for numeric calculations.

All rays have a position and a direction, and each concrete type may carry additional information.

RayTracing.BasicRay — Type

BasicRay{ND, T} <: AbstractRay{ND, T}

A basic implementation of a ray in ND dimensions with arithmetic type T.

Only stores a current position and a direction.

RayTracing.BasicRay — Method

BasicRay{ND, T}(pos, dir) where {ND, T}
BasicRay(pos, dir)

Construct a BasicRay{ND, T} with position pos and direction dir.

With the second signature, ND and T are determined automatically from the arguments.

The direction is normalized before costruction.

Both arguments should be vectors of length ND and element type T. Both are copied so that later changes to the objects passed as arguments don't affect the constructed ray.

RayTracing.RayWithData — Type

RayWithData{ND, T, Data, Ray <: AbstractRay{ND, T}} <: AbstractRayWithData{ND, T, Data}

Wrapper around a ray type, R, to hold arbitrary data of type Data.

The data can be retrieved with data. If you want the data to be mutable, use a mutable Data type or use RayWithMutableData.

RayTracing.RayWithData — Method

RayWithData(ray, data)
RayWithData(pos, dir, data)

Construct a RayWithData with the given data and ray, or with the given data and a BasicRay constructed from pos and dir.

RayTracing.RayWithMutableData — Type

RayWithMutableData{ND, T, Data, Ray <: AbstractRay{ND, T}} <: AbstractRayWithData{ND, T, Data}

Wrapper around a ray type, R, to hold arbitrary mutable data of type Data.

The data can be retrieved with data and assigned to with data!. If you don't need to assign to the data field, use RayWithData instead.

RayTracing.RayWithMutableData — Method

RayWithMutableData(ray, data)
RayWithMutableData(pos, dir, data)

Construct a RayWithMutableData with the given data and ray, or with the given data and a BasicRay constructed from pos and dir.

RayTracing.StateRecorder — Type

StateRecorder(ray, state_func=deepcopy, state_type=typeof(ray))

Construct a StateRecorder with the appropriate type parameters.

state_func must return values of state_type, which are the values to be recorded.

See also StepRecorder.

RayTracing.StateRecorder — Type

StateRecorder{ND, T, State, StateFunc <: Function, R <: AbstractRay{ND, T}} <: AbstractRayWithData{ND, T, State}

Wrapper around a ray type, R, that records the state of the ray after each advance! call.

The state of the ray is a value of type State returned from a function of type StateFunc that takes the ray as argument.

RayTracing.StepRecorder — Type

StepRecorder{ND, T, R <: AbstractRay{ND, T}}

Wrapper around a ray type, R, that records the trajectory of the ray.

A new point is added to the trajectory after every call to advance!. The trajectory can be retrieved by data.

RayTracing.AbstractRayWithData — Type

abstract type AbstractRayWithData{ND, T, Data} <: AbstractRay{ND, T}

A ray that can hold additional data of type Data.

See also: data, RayWithData, RayWithMutableData.

Ray Sources

RayTracing.BackwardCameraRays — Type

BackwardCameraRays{T} <: AbstractMatrix{BasicRay{3, T}}

A rectangular grid of rays starting from a single point.

If used for backward ray tracing, each ray represents a pixel on a camera.

The rays are of type BasicRay{3,T}, and they are evaluated lazily on each call to getindex.

RayTracing.BackwardCameraRays — Method

BackwardCameraRays{T}(; kwargs...)

Construct BackwardCameraRays{T} from keyword arguments describing the camera as in Wikipedia: Calculate rays for rectangular viewport.

Keyword arguments:

origin: The point where the camera is located.

dir: The direction the camera is pointing in. Can't be used together with target.

target: A point the camera is pointing to. Can't be used together with dir.

up: The "up" direction of the camera (controls the rotation of the camera, can be used together with roll). Defaults to [0, 0, 1].

roll: Angle of rotation of the camera around dir, measured with respect to the up direction. Defaults to 0, meaning no rotation.

horizontal_aperture: The angle of aperture of the horizontal dimension of the image (in radians). The angle of aperture for the vertical dimension is determined by the aspect ratio given by image_size.

size: The size of the image, in pixels. Must be a (height, width) tuple.

Reflection Modes

RayTracing.ReflectionMode — Type

abstract type ReflectionMode

Abstract type to describe the way that rays bounce (reflect) on a surface.

Objects of this type, when passed to reflect!, determine how the ray is mutated (i.e. how a new direction is chosen, and what other data is mutated).

RayTracing.reflect! — Function

reflect!(ray::AbstractRay, mode::ReflectionMode, geometric_data, material_data)

Reflect ray on the surface described by mode, geometric_data, and material_data.

The type and meaning of geometric_data and material_data is defined by each type of ReflectionMode.

RayTracing.CombinedReflectionModes — Type

CombinedReflectionModes{Modes <: Tuple{Vararg{ReflectionMode}}} <: ReflectionMode

A ReflectionMode that combines the effects of different reflection modes.

Calling reflect! with this mode results in one call to reflect! with each contained mode.

Example

julia> struct Mode1 <: RayTracing.ReflectionMode end

julia> struct Mode2 <: RayTracing.ReflectionMode end

julia> RayTracing.reflect!(ray, mode::Mode1, geometric_data, material_data) = println(material_data)

julia> RayTracing.reflect!(ray, mode::Mode2, geometric_data, material_data) = println(geometric_data)

julia> combined_mode = CombinedReflectionModes(Mode1(), Mode2())
CombinedReflectionModes{Tuple{Mode1, Mode2}}((Mode1(), Mode2()))

julia> reflect!(:some_ray, combined_mode, "some geometric data", "some material data")
some material data
some geometric data

RayTracing.SpecularReflection — Type

SpecularReflection <: ReflectionMode

This ReflectionMode inverts rays with respect to a surface, mutating only the ray's direction.

RayTracing.SetMaterialData — Type

SetMaterialData <: ReflectionMode

A ReflectionMode that sets the ray's data to the material data passed to reflect!.

This reflection mode should be used with RayWithMutableData or other ray type implementing a data! method.

Example

julia> ray = RayWithMutableData([0,0,0.], [1,1,1.], 5);

julia> data(ray)
5

julia> reflect!(ray, SetMaterialData(), 0, 10);

julia> data(ray)
10

RayTracing.MaterialDataOperation — Type

MaterialDataOperation <: ReflectionMode

A ReflectionMode that applies some arbitrary operation between the ray's data and the material data passed to reflect!.

The result of the operation is set as the ray's data with data! (see RayWithMutableData).

Example

julia> ray = RayWithMutableData([0,0,0.], [1,1,1.], 5);

julia> data(ray)
5

julia> reflect!(ray, MaterialDataOperation(+), 0, 10);

julia> data(ray)
15

Scene Types

RayTracing.AbstractScene — Type

abstract type AbstractScene

A set of objects that rays can bounce on.

Includes information about the geometry of the objects and also about surface properties.

See also: SceneGeometry.

RayTracing.SceneGeometry — Type

abstract type SceneGeometry

A geometric description of a set of objects.

Similar to AbstractScene, but only contains geometric information. Scenes can be created by adding surface information to existing scene geometries.

RayTracing.BallPit — Type

BallPit{ND, T} <: SceneGeometry

A set of same-size spheres in ND dimensions with arithmetic type T.

The arguments for the constructor are:

r::T, the radius of the spheres.

centers::Vector{Point{ND, T}}, the centers of the spheres.

RayTracing.GeometryWithData — Type

GeometryWithData{GeometryType <: SceneGeometry, MappingType} <: AbstractScene

A scene that wraps a SceneGeometry and adds a mapping from surface id to material data.

Scene Interface

RayTracing.next_intersection — Function

next_intersection(ray, scene::AbstractScene) -> distance, surface_id
next_intersection(ray, scene::SceneGeometry) -> distance, surface_id

Calculate the next intersection between ray and any object from scene.

Return the distance to the intersection and an id of the corresponding surface. The type and meaning of the id is defined by the conrete type of scene.

If there are no intersections, the returned distance is infinity and the surface id is unspecified.

RayTracing.get_surface_data — Function

get_surface_data(scene::AbstractScene, surface_id, pos) -> reflection_mode, geometric_data, material_data
get_surface_data(scene::SceneGeometry, surface_id, pos) -> geometric_data

Return surface data for a surface of scene identified by surface_id (as returned by next_intersection) at point pos.

Scene geometries only return data about the geometry of the surface. AbstractScenes return data about geometry as well as material properties and reflection mode.

Tracing Interface

RayTracing.step! — Method

step!(ray, scene)

Advance ray on scene by one step.

This involves finding the next intersection, advancing, and reflecting.

Return :ok if the step was completed, :no_intersection if there are no intersections, or any other Symbol if the step cannot be completed for other reasons.

RayTracing.trace! — Method

trace!(ray, scene; kwargs...)

Trace a ray on a scene.

The ray is evolved by step!ing until there are no more intersections, or an ending condition is met.

Ending conditions are specified by keyword arguments. The default values are such that tracing only stops when there are no more intersections.

The return value is a Symbol indicating why tracing stopped. Possible values are:

:max_steps: The number of steps reached the value of the max_steps keyword argument.

:ray_end_condition: The function passed to the keyword argument ray_end_condition evaluated to true.

Any Symbol other that :ok returned by step!.

Keyword Arguments

max_steps: Stop tracing after this number of steps. Defaults to Inf.

ray_end_condition: A function that takes a ray and returns a boolean. Stop tracing when the return value is true. Defaults to Returns(false).