Base Structs and functions of the program

Raytracing.HDRimage — Type

HDRimage(
    width::Int64
    height::Int64
    rgb_m::Array{RGB{Float32}} = fill(RGB(0.0, 0.0, 0.0), (width*height,))
    )

Define a image in the format 2D High-Dynamic-Range.

Arguments

width::Float64 : width pixel number of the image

height::Float64 : height pixel number of the image

rgb_m::Array{RGB{Float32}} : linearized color matrix; the first element is the one in the bottom-left of the matrix, then the line is read left-to-right and going to the upper row.

Raytracing.to_RGB — Function

to_RGB(r::Int64, g::Int64, b::Int64) :: RGB{Float32}

Return the RGB color with values inside the [0,1] interval.

Raytracing.get_matrix — Function

get_matrix(img::HDRimage) :: Matrix{RGB{Float32}}

Return the color matrix of the input img. The order of the pixel as they are stored in the HDRimage format is corrected in order to get the "natural" pixel matrix.

See also: HDRimage

Raytracing.Parameters — Type

Parameters(
    infile::String, 
    outfile::String,
    a::Float64 = 0.18,
    γ::Float64 = 1.0
    )

Parameters passed from command line for the tone mapping.

Arguments

infile::String : input file name (must be a pfm)

outfile::String : output file name (must be a png)

a::Float64 : parameter a for luminosity correction

γ::Float64 : parameter γ for screen correction

See also: tone_mapping

Raytracing.Point — Type

Point(x::Float64, y::Float64, z::Float64)

A point in 3D space.

Constructors

Point() = new(0., 0. ,0.)

Point(x, y, z) = new(x, y, z)

Point(v::SVector{4, Float64}) = new(v[1], v[2], v[3])

Raytracing.Vec — Type

Vec(x::Float64, y::Float64, z::Float64)

A 3D Vector

Constructors

Vec() = new(0., 0. ,0.)

Vec(x, y, z) = new(x, y, z)

Vec(P::Point) = new(P.x, P.y, P.z)

Vec(v::SVector{4, Float64}) = new(v[1], v[2], v[3])

Vec(N::Normal) = Vec(N.x, N.y, N.z)

See also: Normal

Raytracing.Normal — Type

Normal(x::Float64, y::Float64, z::Float64)

A normal vector in 3D space, you can give three components and its struct normalize them.

Constructors

Normal(x,y,z) = new(x, y ,z)

Normal(v::Vec) = new(v[1]/m, v[2]/m, v[3]/m)

Normal(v::Vector{Float64}) = new(v[1]/m, v[2]/m, v[3]/m)

Normal(v::SVector{4,Float64}) = new(v[1]/m, v[2]/m, v[3]/m)

(m indicates the norm of the input vector)

See also: Vec

Raytracing.Transformation — Type

Transformation(M::SMatrix{4,4,Float64}, invM::SMatrix{4,4,Float64})

Contain two matrices 4x4 of Float64, one the inverse of the other. It's used to implement rotations, scaling and translations in 3D space with homogenous formalism.

NOTE: It does not check if invM is the inverse matrix of M, for computational efficiency purposes! In order to do that, looks at is_consistent(T::Transformation) function.

Constructors

Transformation(m, invm) = new(m, invm)

Transformation() = new( SMatrix{4,4}( Diagonal(ones(4)) ), SMatrix{4,4}( Diagonal(ones(4)) ) )

See also: is_consistent

Raytracing.Ray — Type

Ray(
    origin::Point,
    dir::Vec,
    tmin::Float64 = 1e-5,
    tmax::Float64 = Inf,
    depth::Int64 = 0,
    )

A ray of light propagating in space.

Arguments

origin::Point : origin of the ray

dir::Vec : 3D direction along which this ray propagates

tmin::Float64 : minimum distance travelled by the ray is this number times dir

tmax::Float64 : maximum distance travelled by the ray is this number times dir

depth::Int64 : number of times this ray was reflected/refracted

See also: Point, Vec

Raytracing.at — Function

at(r::Ray, t::Float64) :: Point

Compute the point along the ray's path at some distance from the origin.

Return a Point object representing the point in 3D space whose distance from the ray's origin is equal to t, measured in units of the length of Vec.dir.

See also: Ray, Point, Vec

Raytracing.Camera — Type

abstract type Camera end

An abstract type with the following concrete sub-types, defining different types of perspective projections:

OrthogonalCamera

PerspectiveCamera

Raytracing.fire_ray — Function

fire_ray(Ocam::OrthogonalCamera, u::Float64, v::Float64) :: Ray
fire_ray(Pcam::PerspectiveCamera, u::Float64, v::Float64) :: Ray
fire_ray(
        ImTr::ImageTracer, 
        col::Int64, row::Int64, 
        u_px::Float64=0.5, v_px::Float64=0.5
    ) :: Ray

Shoot one light Ray through the pixel (col, row) of ImTr.img image. The parameters (col, row) are measured in a diffetent way compared to the HDRimage struct: here, the bottom left corner is placed at (0, 0). The following diagram shows the convenction for an image with dimensions (w,h):

| (h,0)  (h,1)  (h,2)  ...  (h,w) |
|  ...    ...    ...   ...   ...  |
| (1,0)  (1,1)  (1,2)  ...  (1,w) |
| (0,0)  (0,1)  (0,2)  ...  (0,w) |

The optional values u_px and v_px specify where the ray should cross the pixel; the convenction for their values are represented in the following diagram as (u_px, v_px):

(0, 1)                          (1, 1)
    +------------------------------+
    |                              |
    |                              |
    |                              |
    +------------------------------+
(0, 0)                          (1, 0)

See also: OrthogonalCamera, PerspectiveCamera, Ray, HDRimage, ImageTracer

Raytracing.ImageTracer — Type

ImageTracer(
    img::HDRimage,
    cam::Camera,
    samples_per_side::Int64 = 0,
    pcg::PCG = PCG()
    )

Implement the "screen" of the observer.

Trace an image by shooting light rays through each of its pixels.

Arguments

img::HDRimage : the image that will be rendered (required)

cam::Camera : camera type of the observer (required)

samples_per_side::Int64 = 0 : if it is larger than zero, stratified sampling will be applied to each pixel in the image, using the random number generator pcg; if not, antialiasing will be ignored in fire_all_rays!

pcg::PCG = PCG() : PCG random number generator

See also: HDRimage,Camera, PCG, fire_ray, fire_all_rays!

Raytracing.fire_all_rays! — Function

fire_all_rays!(
        ImTr::ImageTracer, 
        func::Function, 
        callback::Union{Nothing, Function} = nothing,
        callback_time_s::Float64 = 2.,
        callback_kwargs::String
        )

Shoot several light rays crossing each of the pixels in the ImTr.img image.

For each pixel in the HDRimage object fire one Ray, and pass it to the function func, which must:

accept a Ray as its only parameter

return a RGB{Float32} color instance telling the color to assign to that pixel in the image.

If callback is not nothing, it must be a function accepting at least two parameters named col and row. This function is called periodically during the rendering, and the two mandatory arguments are the row and column number of the last pixel that has been traced.

Pay Attention: Both the row and column are increased by one starting from zero: first the row and then the column.

The time between two consecutive calls to the callback can be tuned using the parameter callback_time_s. Any keyword argument passed to fire_all_rays is passed to the callback.

See also: Ray, HDRimage, ImageTracer

Raytracing.Pigment — Type

abstract type Pigment end

This abstract class represents a pigment, i.e., a function that associates a color with each point on a parametric surface (u,v). Call the function get_color to retrieve the color of the surface given a Vec2d object.

The concrete sub-types of this abstract class are:

UniformPigment

CheckeredPigment

ImagePigment

See also: Vec2d, get_color

Raytracing.BRDF — Type

abstract type BRDF end

An abstract class representing a Bidirectional Reflectance Distribution Function. The concrete sub-types of this abstract class are:

DiffuseBRDF

SpecularBRDF

Raytracing.Material — Type

Material(
    brdf::BRDF = DiffuseBRDF(),
    emitted_radiance::Pigment = UniformPigment()
)

A struct representing a material.

See also: BRDF, DiffuseBRDF, Pigment, UniformPigment

Raytracing.Shape — Type

abstract type Shape end

An abstract type with the following concrete sub-types, defining different types of shapes that can be created:

Sphere

Plane

Raytracing.Vec2d — Type

Vec2d(u::Float64, v::Float64)

A 2D vector used to represent a point on a surface. The fields are named u and v to distinguish them from the usual 3D coordinates x, y, z.

Raytracing.HitRecord — Type

HitRecord(
    world_point::Point,
    normal::Normal,
    surface_point::Vec2d,
    t::Float64,
    ray::Ray,
    shape::Union{Shape, Nothing} = nothing
    )

A struct holding information about a ray-shape intersection.

Arguments

world_point::Point: world coordinates of the hit point

normal::Normal: orientation of the normal to the surface where the hit happened

surface_point::Vec2d : position of the hit point on the surface of the object

t::Float64 : distance from the origin of the ray where the hit happened

ray::Ray : the ray that hit the surface

shape::Union{Shape, Nothing}: shape on which the hit happened, or nothing if no intersection happened

See also: Point, Normal, Vec2d Ray, Shape

Raytracing.PointLight — Type

PointLight(p::Point, c::Color, r::Float64 = 0.0)

A point light (used by the point-light renderer). This class holds information about a point light (a Dirac's delta in the rendering equation).

Arguments

position::Point : position of the point light in 3D space

color::RGB{Float32} : color of the point light

linear_radius::Float64: if non-zero, this «linear radius» r is used to compute the solid angle subtended by the light at a given distance d through the formula $(r / d)^2$.

See also: Point, PointLightRenderer

Raytracing.World — Type

World(
    shapes::Array{Shape} = Array{Shape,1}(),
    point_lights::Array{PointLight} = Array{PointLight,1}()
)

A struct holding a list of shapes, which make a «world».

You can add shapes to a world using add_shape!, and call ray_intersection to check whether a light ray intersects any of the shapes in the world.

For the PointLightRenderer algorithm, you can also add point-lights source using add_light!, and world will keep a list of all of them.

See also: Shape, add_shape!, PointLight, add_light, PointLightRenderer ray_intersection, Ray

Raytracing.add_shape! — Function

add_shape!(world::World, shape::Shape)

Append a new shape to the given world.

See also: Shape, World

Raytracing.add_light! — Function

add_light!(world::World, pointlight::PointLight)

Append a new pointlight to the given world.

See also: PointLight, World

Raytracing.PCG — Type

 PCG(state::UInt64 = UInt64(42), inc::UInt64 = UInt64(54))

A mutable struct for the PCG Uniform Pseudo-random Number Generator.

Parameters

state::UInt64 = UInt64(42) : initial state number

inc::UInt64 = UInt64(54) : initial sequence number

References

Melissa E. O’Neill (2014), "PCG: A Family of Simple Fast Space-Efficient Statistically Good Algorithms for Random Number Generation"

See also: random(pcg::PCG, ::Type{UInt32}), random(pcg::PCG, ::Type{Float64})

Raytracing.random — Function

 random(pcg::PCG, ::Type{UInt32}) :: UInt32

Return a new random UInt32 number and advance PCG's internal state.

This function is based on the paper of Melissa E. O’Neill (2014), where the Permuted Congruential Generator (PCG) family of random number generators is defined and explained.

References

Melissa E. O’Neill (2014), "PCG: A Family of Simple Fast Space-Efficient Statistically Good Algorithms for Random Number Generation"

See also: random(pcg::PCG, ::Type{Float64}), PCG

 random(pcg::PCG, ::Type{Float64}) :: Float64

Returns a Float64 random number inside [0,1] interval obtained with a PCG Uniform Pseudo-random Number Generator.

It calls the random(pcg::PCG, ::Type{UInt32}) function, which returns a UInt32 random number, and divides it with typemax(UInt32).

See also: random(pcg::PCG, ::Type{UInt32}), PCG

 random(pcg::PCG) :: Float64

Returns a Float64 random number inside [0,1] interval obtained with a PCG Uniform Pseudo-random Number Generator.

It calls the random(pcg::PCG, ::Type{Float64}) function.

See also: random(pcg::PCG, ::Type{Float64}), random(pcg::PCG, ::Type{UInt32}), PCG

Raytracing.create_onb_from_z — Function

create_onb_from_z(normal::Union{Vec, Normal}) :: (Normal, Normal, Normal)

Return an orthonormal base of 3 Normals with the z-axes (i.e. the third Normal returned) parallel to the input Vec/ Normal.

The implementation of this function is based on the paper of Duff et al. (2017), which improved the already efficient work made by Frisvad (2012).

References

Duff et al. (2017), "Building an Orthonormal Basis, Revisited"

Frisvad (2012), "Building an Orthonormal Basis from a 3D Unit Vector Without Normalization"

See also: Vec, Normal

Raytracing.scatter_ray — Function

scatter_ray(
    ::Type{DiffuseBRDF},
    pcg::PCG, 
    incoming_dir::Vec, 
    interaction_point::Point, 
    normal::Normal, 
    depth::Int64,
    ) :: Ray

Return a Ray scattered by a material with a DiffuseBRDF.

A DiffuseBRDF has a uniform BRDF, i.e. $f_r(\mathbf{x}, \mathbf{\Psi}\rightarrow\mathbf{\Theta}) = \rho_d / \pi$; the importance sampling for the PathTracer algorithm use consequently the following PDF:

\[p(\omega) \propto f_r(\mathbf{x}, \mathbf{\Psi}\rightarrow\mathbf{\Theta}) \, \cos(\vartheta) = \frac{\rho_d}{\pi} \, \cos(\vartheta) \propto \cos(\vartheta)\]

\[ \Rightarrow \quad p(\omega) = \frac{\cos(\vartheta)}{\pi} \quad \Rightarrow \quad p(\vartheta ,\varphi) = \frac{\cos(\vartheta) \, \sin(\vartheta) }{2\pi}\]

\[\Rightarrow \quad \begin{aligned} &p(\vartheta) = 2 \, \cos(\vartheta) \, \sin(\vartheta) \\ &p(\varphi | \vartheta) = \frac{1}{2 \pi} \end{aligned}\]

See also: DiffuseBRDF, Ray, Vec, Point, Normal, PCG

scatter_ray(
    ::Type{SpecularBRDF},
    pcg::PCG, 
    incoming_dir::Vec, 
    interaction_point::Point, 
    normal::Normal, 
    depth::Int64,
    ) :: Ray

Return a Ray scattered by a material with a SpecularBRDF.

A SpecularBRDF has a Dirac delta BRDF, i.e.:

\[f_r(\mathbf{x}, \mathbf{\Psi} \rightarrow \mathbf{\Theta}) \propto \frac{\delta(\sin^2\theta_r - \sin^2\theta) \, \delta(\psi_r \pm \pi - \psi)}{\cos\theta},\]

The importance sampling for the PathTracer algorithm use consequently the following PDF:

\[p(\omega) \propto f_r(\mathbf{x}, \mathbf{\Psi}\rightarrow\mathbf{\Theta}) \, \cos(\vartheta) \propto \frac{1}{\cos(\vartheta)} \, \cos(\vartheta) \propto cost\]

See also: SpecularBRDF, Ray, Vec, Point, Normal, PCG