Demo

Raytracing.demo — Function

demo(
      camera_type::String = "per",
	camera_position::Point = Point(-1.,0.,0.), 
	algorithm::String = "flat",
      α::Float64 = 0., 
      width::Int64 = 640, 
      height::Int64 = 480, 
      pfm_output::String = "demo.pfm", 
      png_output::String = "demo.png",
	bool_print::Bool = true,
	bool_savepfm::Bool = true,
	world_type::String = "A",
	init_state::Int64 = 45,
	init_seq::Int64 = 54,
	samples_per_pixel::Int64 = 0
      )

Creates a demo image with the specified options.

There are two possible demo image "world" to be rendered, specified through the input string type.

The type=="A" demo image world consist in a set of 10 spheres of equal radius 0.1: 8 brown spheres are placed at the verteces of a cube of side 1.0, one green-purple checked sphere is in the center of the lower cube face and another multi-colored sphere is in the center of the left cube face.

The type=="B" demo image world consists in a checked x-y plane, a blue opaque sphere, a red reflecting sphere, and a green oblique reflecting plane, all inside a giant emetting sphere.

The type=="C" demo image world consists in... discover yourself!

The creation of the demo image has the objective to check the correct behaviour of the rendering software, specifically the orientation upside-down and left-right.

Arguments

camera_type::String = "per" : set the perspective projection view:

- `camera_type=="per"` -> set [`PerspectiveCamera`](@ref)  (default value)
	- `camera_type=="ort"`  -> set [`OrthogonalCamera`](@ref)

camera_position::Point = Point(-1.,0.,0.) : set the point of observation in (X,Y,Z`) coordinates

algorithm::String = "flat" : algorithm to be used in the rendered:

algorithm=="onoff" -> OnOffRenderer algorithm

algorithm=="flat" -> FlatRenderer algorithm (default value)

algorithm=="pathtracing" -> PathTracer algorithm

algorithm=="pointlight" -> PointLightRenderer algorithm

α::Float64 = 0. : angle of rotation IN RADIANTS, relative to the vertical (i.e. z) axis, of the view direction

width::Int64 = 640 and height::Int64 = 480 : pixel dimensions of the demo image

pfm_output::String = "demo.pfm" : name of the output pfm file

png_output::String = "demo.png" : name of the output LDR file

bool_print::Bool = true : specifies if the WIP messages of the demo function should be printed or not (useful option for demo_animation)

bool_savepfm::Bool = true : bool that specifies if the pfm file should be saved or not (useful option for demo_animation)

world_type::String = "A" : specifies the type of world to be rendered ("A", "B" or "C")

init_state::Int64 = 45 : initial state of the PCG random number generator

init_seq::Int64 = 54 : initial sequence of the PCG random number generator

samples_per_pixel::Int64 = 0 : number of rays per pixel to be used (antialiasing)

See also: Point ,OnOffRenderer, FlatRenderer, PathTracer, demo_animation

Raytracing.first_world — Function

first_world() :: World

Render the first world (identified with the string "A").

This world consists in a set of 10 spheres of equal radius 0.1: 8 brown spheres are placed at the verteces of a cube of side 1.0, one green-purple checked sphere is in the center of the lower cube face and another multi-colored sphere is in the center of the left cube face.

See also: World, demo, demo_animation

Raytracing.second_world — Function

second_world() :: World

Render the second world (identified with the string "B").

This world consists in a checked x-y plane, a blue opaque sphere, a red reflecting sphere, and a green oblique reflecting plane, all inside a giant emetting sphere.

See also: World, demo, demo_animation

Raytracing.select_world — Function

select_world(type_world::String) ::Function

Select which demo world is used

Raytracing.parse_demo_settings — Function

parse_demo_settings(dict::Dict{String, Any}) 
    :: (
        String, Point, String, Float64, Int64, Int64, String, String,
        Bool, Bool, String, Int64, Int64, Int64
        )

Parse a Dict{String, T} where {T} for the demo function.

Input

A `dict::Dict{String, T} where {T}

Returns

A tuple (ct, cp, al, α, w, h, pfm, png, bp, bs, wt, ist, ise, spp) containing the following variables; the corresponding keys are also showed:

ct::String = dict["camera_type"] : set the perspective projection view:

ct=="per" -> set PerspectiveCamera (default value)

ct=="ort" -> set OrthogonalCamera

cp::String = dict["camera_position"] : "X,Y,Z" coordinates of the choosen observation point of view

al::String = dict["algorithm"] : algorithm to be used in the rendered:

al=="onoff" -> OnOffRenderer algorithm

al=="flat" -> FlatRenderer algorithm (default value)

al=="pathtracing" -> PathTracer algorithm

algorithm=="pointlight" -> PointLightRenderer algorithm

α::String = dict["alpha"] : choosen angle of rotation respect to vertical (i.e. z) axis

w::Int64 = dict["width"] : number of pixels on the horizontal axis to be rendered

h::Int64 = dict["height"] : number of pixels on the vertical axis to be rendered

pfm::String = dict["set_pfm_name"] : output pfm filename

png::String = dict["setpngname"]` : output LDR filename

bp::Bool = dict["bool_print"] : if true, WIP message of demo function are printed (otherwise no)

bs::Bool = dict["bool_savepfm"] : if true, demo function saves the pfm file to disk

wt::String = dict["world_type"] : type of the world to be rendered

ist::Int64 = dict["init_state"] : initial state of the PCG generator

ise::Int64 = dict["init_seq"] : initial sequence of the PCG generator

spp::Int64 = dict["samples_per_pixel"] : number of ray to be generated for each pixel

See also: demo, Point, PCG