Demo

Raytracing.demo_animation — Function

demo_animation( 
		camera_type::String = "per",
		algorithm::String = "flat",
    		width::Int64 = 200, 
    		height::Int64 = 150,
		world_type::String = "A",
   		anim_output::String = "demo-animation.mp4",
	)

Creates an animation of the demo image with the specified options. It's necessary to have istalled the ffmpeg software to run this function.

This function works following this steps:

creates an hidden directory, called ".wip_animation"; if it already exists, it will be destroyed and recreated.

inside ".wpi_animation", creates 360 png images of the demo image (using the demo function with the specified projection, algorithm and image dims); each image correspons to a frame of the future animation

through the ffmpeg software, the 360 png images are converted into the animation mp4 file, and saved in the main directory

the ".wpi_animation" directory and all the png images inside it are destroyed

Arguments

camera_type::String = "per" : set the perspective projection view:

- `camera_type=="per"` -> set [`PerspectiveCamera`](@ref)  (default value)
	- `camera_type=="ort"`  -> set [`OrthogonalCamera`](@ref)

algorithm::String = "flat" : algorithm to be used in the rendered:

algorithm=="onoff" -> OnOffRenderer algorithm

algorithm=="flat" -> FlatRenderer algorithm (default value)

algorithm=="pathtracing" -> PathTracer algorithm

algorithm=="pointlight" -> PointLightRenderer algorithm

width::Int64 = 640 and height::Int64 = 480 : pixel dimensions of the demo image

world_type::String = "A" : specifies the type of world to be rendered ("A", "B" or "C")

anim_output::String = "demo-animation.mp4" : name of the output animation file

samples_per_pixel::Int64 = 0 : number of rays per pixel to be used (antialiasing)

See also: OnOffRenderer, FlatRenderer, PathTracer, demo

Raytracing.parse_demoanimation_settings — Function

parse_demoanimation_settings(dict::Dict{String, T}) where {T}
    :: (String, String, Int64, Int64, String, String, Int64)

Parse a Dict{String, T} where {T} for the demo_animation function.

Input

A dict::Dict{String, T} where {T}

Returns

A tuple (ct, al, w, h, wt, anim, spp) containing the following variables; the corresponding keys are also showed:

ct::String = dict["camera_type"] : set the perspective projection view:

- `ct=="per"` -> set [`PerspectiveCamera`](@ref)  (default value)
	- `ct=="ort"`  -> set [`OrthogonalCamera`](@ref)

al::String = dict["algorithm"] : algorithm to be used in the rendered:

al=="onoff" -> OnOffRenderer algorithm

al=="flat" -> FlatRenderer algorithm (default value)

al=="pathtracing" -> PathTracer algorithm

algorithm=="pointlight" -> PointLightRenderer algorithm

w::Int64 = dict["width"] : number of pixels on the horizontal axis to be rendered

h::Int64 = dict["height"] : width and height of the rendered image

wt::String = dict["world_type"] : type of the world to be rendered

anim::String = dict["set_anim_name"] : output animation name

spp::Int64 = dict["samples_per_pixel"] : number of ray to be generated for each pixel

See also: demo_animation, demo