Scene Files

Basic functions

Raytracing.Interpreter.isdigit — Function

isdigit(a::String) :: Bool

Return true if a is a sigle digit, false otherwise.

Raytracing.Interpreter.isdecimal — Function

isdecimal(a::String) :: Bool

Return true if a is an integer number, false otherwise.

Raytracing.Interpreter.isalpha — Function

isalpha(a::String) :: Bool

Return true if a is a string made only of the 26 english letters (capitalized or not) and/or the underscore symbol "_" , false otherwise.

Raytracing.Interpreter.isalnum — Function

isalnum(a::String) :: Bool

Return true if a is a string made only of the 26 english letters (capitalized or not), the underscore symbol "_" and the 10 basic digits, false otherwise.

Raytracing.Interpreter.close_bracket — Function

 closed_bracket(open::String) :: String

Given in input a string of an open braket (["{", "[", "(", "<"]) return the corresponding closed one (["}", "]", ")", ">"]).

SourceLocation and Tokens

Raytracing.Interpreter.SourceLocation — Type

SourceLocation(file_name::String, line_num::Int64, col_num::Int64)

A specific position in a source file.

Arguments

file_name::String : the name of the file, or the empty string if there is no file associated with this location (e.g., because the source code was provided as a memory stream, or through a network connection)

line_num::Int64 : number of the line (starting from 1)

col_num::Int64 : number of the column (starting from 1)

Raytracing.Interpreter.KeywordEnum — Type

 @enum KeywordEnum

Enumeration for all the possible keywords recognized by the lexer:

|:-----------------:|:-----------------:|:----------------------:|
| NEW = 1           | PIGMENT = 20      | TRANSFORMATION = 40    |
| FLOAT = 2         | UNIFORM = 21      | IDENTITY = 41          |
| STRING = 3        | CHECKERED = 22    | TRANSLATION = 42       |
| VECTOR = 4        | IMAGE = 23        | ROTATION_X = 43        |
| COLOR = 5         |                   | ROTATION_Y = 44        |
| MATERIAL = 6      |                   | ROTATION_Z = 45        |
| POINTLIGHT = 7    |                   | SCALING = 46           |
|                   |                   |                        |
|:-----------------:|:-----------------:|:----------------------:|
| BRDFS = 10        | CAMERA = 30       | BOOL = 50              |
| DIFFUSE = 11      | ORTHOGONAL = 31   | TRUE = 51              |
| SPECULAR = 12     | PERSPECTIVE = 32  | FALSE = 52             |
|                   |                   |                        |
|                   |                   |                        |
|:-----------------:|:-----------------:|:----------------------:|
| PLANE = 61        | PRINT = 71        |                        |
| SPHERE = 62       | ASSERT = 72       |                        |
| CUBE = 63         |                   |                        |
| TRIANGLE = 64     |                   |                        |
|                   |                   |                        |
|:-----------------:|:-----------------:|:----------------------:|

Raytracing.Interpreter.Token — Type

Token(
      location::SourceLocation,
      value::Union{  
           KeywordToken, 
           IdentifierToken, 
           StringToken,
           LiteralNumberToken,
           SymbolToken, 
           StopToken}
      )

A lexical token, used when parsing a scene file.

Arguments

location::SourceLocation: location of the last char read

value : one of the basic 6 token types:

KeywordToken

IdentifierToken

StringToken

LiteralNumberToken

SymbolToken

StopToken

See also: SourceLocation

Raytracing.Interpreter.KeywordToken — Type

KeywordToken(keyword::KeywordEnum)

A token containing a keyword of the Photorealistic Object Applications Language.

See also: KeywordEnum

Raytracing.Interpreter.IdentifierToken — Type

IdentifierToken(identifier::String)

A token containing an identifier, i.e. a name of a variable.

Raytracing.Interpreter.StringToken — Type

 StringToken(string::String)

A token containing a literal string, i.e. a sentence placed inside two double quotes ("...") symbols.

Raytracing.Interpreter.LiteralNumberToken — Type

LiteralNumberToken(number::Float64)

A token containing a literal number.

Raytracing.Interpreter.SymbolToken — Type

SymbolToken(symbol::String)

A token containing a recognised symbol by the Photorealistic Object Applications Language.

Raytracing.Interpreter.StopToken — Type

StopToken()

A token signalling the end of a file.

Raytracing.Interpreter.GrammarError — Type

 GrammarError <: Exception(
      location::SourceLocation
      message::String
 )

An error found by the lexer/parser while reading a scene file.

Arguments

location::SourceLocation : location of the last char read

message::String : a user-frendly error message

See also: SourceLocation

Parsing and readingTokens

Raytracing.Interpreter.InputStream — Type

InputStream(
    stream::IO,
    location::SourceLocation,
    saved_char::String,
    saved_location::SourceLocation,
    tabulations::Int64,
    saved_token::Union{Token, Nothing},
)

A high-level wrapper around a stream, used to parse scene files This class implements a wrapper around a stream, with the following additional capabilities:

It tracks the line number and column number;

It permits to "un-read" characters and tokens.

Arguments

stream::IO : stream to read from

location::SourceLocation : location of the last char read

saved_char::String : the last char read

saved_location::SourceLocation : location where saved_char is in the file

tabulations::Int64: number of space a tab command gives

saved_token::Union{Token, Nothing} : the last token found

Constructors

InputStream( stream::IO, filename::String = "", tabulations::Int64 = 8 ) = new( stream, SourceLocation(filename, 1, 1), "", SourceLocation(file_name, 1, 1), tabulations, nothing )

InputStream( s::IO, l::SourceLocation, sc::String, sl::SourceLocation, t::Int64, st::Union{Token, Nothing} ) = new(s,l,sc,sl,t,st)

See also: SourceLocation, Token

Raytracing.Interpreter.update_pos — Function

update_pos(inputstream::InputStream, ch::String)

Update location after having read ch from the stream.

See also: SourceLocation

Raytracing.Interpreter.read_char — Function

 read_char(inputstream::InputStream) :: String

Read a new character from the stream. Calls internally update_pos.

See also: InputStream, unread_char

Raytracing.Interpreter.unread_char — Function

 unread_char(inputstream::InputStream, ch::String)

Push a character back to the stream.

See also: InputStream, read_char

Raytracing.Interpreter.skip_whitespaces_and_comments — Function

 skip_whitespaces_and_comments(inputstream::InputStream)

Keep reading characters until a non-whitespace/non-comment character is found. Calls internally read_char and unread_char, and it's used inside the main function read_token.

See also: InputStream

Raytracing.Interpreter.parse_string_token — Function

 parse_string_token(
      inputstream::InputStream,
      token_location::SourceLocation
      ) ::Token(::SourceLocation, ::StringToken)

Parse a string from the given input inputstream and return that string inside a Token(::SourceLocation, ::StringToken) with the given token_location, throwing GrammarError in case of exception. Works calling read_char, and it's used inside the main function read_token.

See also: InputStream, SourceLocation Token, StringToken, GrammarError

Raytracing.Interpreter.parse_float_token — Function

 parse_float_token(
      inputstream::InputStream, 
      first_char::String, 
      token_location::SourceLocation
      ) :: Token{SourceLocation, LiteralNumberToken}

Parse a float from the given input inputstream and return that float inside a Token(::SourceLocation, ::LiteralNumberToken) with the given token_location, throwing GrammarError in case of exception. Works calling read_char and unread_char, and it's used inside the main function read_token.

See also: InputStream, SourceLocation Token, LiteralNumberToken, GrammarError

Raytracing.Interpreter.parse_keyword_or_identifier_token — Function

 parse_keyword_or_identifier_token(
      inputstream::InputStream,
      first_char::String,
      token_location::SourceLocation
      ) ::Union{
                Token(::SourceLocation, ::KeywordToken),
                Token(::SourceLocation, ::IdentifierToken)
                }

Parse a keyword or an identifier from the given input inputstream and return that keyword/identifier inside respectively a Token(::SourceLocation, ::KeyworkdToken) or a Token(::SourceLocation, ::IdentifierToken) with the given token_location, throwing GrammarError in case of exception. Works calling read_char and unread_char, and it's used inside the main function read_token.

See also: InputStream, SourceLocation Token, LiteralNumberToken, GrammarError

Raytracing.Interpreter.read_token — Function

 read_token(inputstream::InputStream) :: Token

Read one of the 6 basic tokens from the stream, raising GrammarError if a lexical error is found. Calls internally the following functions:

skip_whitespaces_and_comments

read_char

isdecimal

isalpha

copy(::SourceLocation)

parse_string_token for StringToken

parse_float_token for LiteralNumberToken

parse_keyword_or_identifier_token for KeywordToken and IdentifierToken

See also: InputStream, Token, SymbolToken, StopToken, GrammarError,

Raytracing.Interpreter.unread_token — Function

unread_token(inputstream::InputStream, token::Token)

Make as if token were never read from inputstream.

See also: InputStream, Token, read_token

Scene and basic parsing scene functions

Raytracing.Interpreter.Scene — Type

 Scene(
      materials::Dict{String, Material} = Dict{String, Material}(),
      world::World = World(),
      camera::Union{Camera, Nothing} = nothing,

      float_variables::Dict{String, Float64} = Dict{String, Float64}(),
      string_variables::Dict{String, String} = Dict{String, String}(),
      bool_variables::Dict{String, Bool} = Dict{String, Bool}(),
      vector_variables::Dict{String,Vec} = Dict{String,Vec}(),
      color_variables::Dict{String,RGB{Float32}} = Dict{String,RGB{Float32}}(),
      pigment_variables::Dict{String,Pigment} = Dict{String,Pigment}(),
      brdf_variables::Dict{String,BRDF} = Dict{String,BRDF}(),
      transformation_variables::Dict{String,Transformation} = Dict{String,Transformation}(),

      variable_names::Set{String} = Set{String}(),
      overridden_variables::Set{String} = Set{String}(),
 )

A scene read from a scene file.

See also: Material, World, Camera, Vec, Pigment, BRDF, Transformation

Raytracing.Interpreter.expect_symbol — Function

 expect_symbol(inputstream::InputStream, symbol::String)
 expect_symbol(inputstream::InputStream, vec_symbol::Vector{String}) :: String

Read a token from inputstream and check that its type is SymbolToken and its value is symbol(first method) or a value inside vec_symbol (second method, and return it), throwing GrammarError otherwise. Call internally read_token.

See also: InputStream, KeywordEnum, SymbolToken

Raytracing.Interpreter.expect_keywords — Function

 expect_keywords(inputstream::InputStream, keywords::Vector{KeywordEnum}) :: KeywordEnum

Read a token from inputstream and check that its type is KeywordToken and its value is one of the keywords in keywords, throwing GrammarError otherwise. Call internally read_token.

See also: InputStream, KeywordEnum, KeywordToken

Raytracing.Interpreter.expect_number — Function

 expect_number(inputstream::InputStream, scene::Scene) :: Float64

Read a token from inputstream and check that its type is LiteralNumberToken (i.e. a number) or IdentifierToken (i.e. a variable defined in scene), throwing GrammarError otherwise. Return the float64-parsed number or the identifier associated float64-parsed number, respectively. Call internally read_token.

See also: InputStream, Scene, LiteralNumberToken, IdentifierToken

Raytracing.Interpreter.expect_bool — Function

 expect_bool(inputstream::InputStream, scene::Scene) :: Bool

Read a token from inputstream and check that its type is KeywordToken or IdentifierToken (i.e. a variable defined in scene), throwing GrammarError otherwise. Return the parsed bool or the identifier associated parsed bool, respectively. Call internally read_token.

See also: InputStream, Scene, KeywordToken, IdentifierToken

Raytracing.Interpreter.expect_string — Function

 expect_string(inputstream::InputStream, scene::Scene) :: String

Read a token from inputstream and check that its type is StringToken, throwing GrammarError otherwise. Return the string associated with the readed StringToken. Call internally read_token.

See also: InputStream, Scene, StringToken,

Raytracing.Interpreter.expect_identifier — Function

 expect_identifier(inputstream::InputStream) :: String

Read a token from inputstream and check that it is an identifier. Return the name of the identifier.

Read a token from inputstream and check that its type is IdentifierToken, throwing GrammarError otherwise. Return the name of the identifier as a String. Call internally read_token.

See also: InputStream, Scene, IdentifierToken,

Raytracing.Interpreter.parse_vector — Function

parse_vector(inputstream::InputStream, scene::Scene) :: Vec

Parse a vector from the given inputstream and return it. Call internally expect_number and expect_symbol.

See also: InputStream, Scene, Vec

Raytracing.Interpreter.parse_color — Function

 parse_color(inputstream::InputStream, scene::Scene) :: RGB{Float32}

Read the color from the given inputstream and return it. Call internally 'expect_symbol' and 'expect_number'.

See also: 'InputStream', 'Scene'

Raytracing.Interpreter.parse_pigment — Function

 parse_pigment(inputstream::InputStream, scene::Scene) :: Pigment

Parse a pigment from the given inputstream and return it.

Call internally the following parsing functions:

expect_keywords

expect_symbol

parse_color

expect_number

expect_string

Call internally the following functions and structs of the program:

UniformPigment

CheckeredPigment

ImagePigment

load_image

See also: InputStream, Scene, Pigment

Raytracing.Interpreter.parse_brdf — Function

 parse_brdf(inputstream::InputStream, scene::Scene) :: BRDF

Parse a BRDF from the given inputstream and return it.

Call internally the following parsing functions:

expect_keywords

expect_symbol

parse_pigment

Call internally the following functions and structs of the program:

DiffuseBRDF

SpecularBRDF

See also: InputStream, Scene, BRDF

Raytracing.Interpreter.parse_material — Function

 parse_material(inputstream::InputStream, scene::Scene) :: (String, Material)

Parse a Material from the given inputstream and return a tuple with the identifier name of the material and the material itself.

Call internally the following parsing functions:

expect_identifier

expect_symbol

parse_brdf

parse_pigment

Call internally the following functions and structs of the program:

Material

See also: InputStream, Scene, Token, Material

Raytracing.Interpreter.parse_transformation — Function

 parse_transformation(inputstream::InputStream, scene::Scene) :: Transformation

Parse a Transformation from the given inputstream and return it.

Call internally the following parsing functions:

expect_keywords

expect_symbol

expect_number

parse_vector

read_token

unread_token

Call internally the following functions and structs of the program:

translation

rotation_x

rotation_y

rotation_z

scaling

See also: InputStream, Scene, Transformation

Raytracing.Interpreter.parse_pointlight — Function

 parse_pointlight(inputstream::InputStream, scene::Scene) :: PointLight

Parse a PointLight from the given inputstream and return it.

Call internally the following parsing functions:

read_token

unread_token

expect_number

expect_symbol

parse_vector

parse_color

See also: InputStream, Scene, PointLight

Raytracing.Interpreter.parse_camera — Function

 parse_camera(inputstream::InputStream, scene::Scene) :: Camera

Parse a Camera from the given inputstream and return it.

Call internally the following parsing functions:

expect_symbol

expect_keywords

expect_number

parse_transformation

Call internally the following functions and structs of the program:

OrthogonalCamera

PerspectiveCamera

See also: InputStream, Scene, Camera

Raytracing.Interpreter.return_token_value — Function

 return_token_value(
      inputstream::InputStream, 
      scene::Scene
      ) :: Union{Float64, Bool, String, Vec, RGB{Float32}, Pigment, BRDF}

Return the value of the token readed from inputstream. If the token is an IdentifierToken, return the value associated with that identifier insdie the given scene. Throws GrammarError if an error occurs. Call internally the following parsing functions:

read_token

unread_token

parse_vector

parse_color

parse_pigment

parse_brdf

See also: Vec, Pigment, BRDF, InputStream, Scene, GrammarError

Raytracing.Interpreter.assert — Function

 assert(inputstream::InputStream, scene::Scene)

Parse an assertion from the given inputstream and return it. Throws AssertionError if the assertion is false, nothing otherwise.

Call internally the following parsing functions:

expect_symbol

expect_keywords

expect_number

expect_string

return_token_value

Examples

assert(1, 1)             # Checks that 1==1
assert(1, 2, "<")        # Checks that 1<2
float var(1.0)           # Define var as a variable with value 1.0
assert(var, 2, "<=")     # Checks that var<=2

See also: InputStream, Scene

Raytracing.Interpreter.parse_scene — Function

 parse_scene(
      inputstream::InputStream, 
      variables::Dict{String, Float64} = Dict()
      ) :: Scene

Read a scene description from the given inputstream and return a Scene object. Throws GrammarError if an error occurs.

Call internally the following parsing functions:

read_token

StopToken

expect_identifier

expect_symbol

expect_number

parse_sphere

parse_plane

parse_camera

parse_material

Call internally the following functions and structs of the program:

add_shape!

add_light!

See also: InputStream, Scene

Parsing Shapes

Raytracing.Interpreter.parse_sphere — Function

 parse_sphere(inputstream::InputStream, scene::Scene) :: Sphere

Parse a Sphere from the given inputstream and return it. Throws GrammarError if the specified Material does not exist.

Call internally the following parsing functions:

expect_symbol

expect_identifier

parse_transformation

See also: InputStream, Scene, Sphere Material

Raytracing.Interpreter.parse_plane — Function

 parse_plane(inputstream::InputStream, scene::Scene) :: Plane

Parse a Plane from the given inputstream and return it. Throws GrammarError if the specified Material does not exist.

Call internally the following parsing functions:

expect_symbol

expect_identifier

parse_transformation

See also: InputStream, Scene, Plane, Material