The render function

Raytracing.check_is_positive — Function

 check_is_positive(string::String="") :: Bool
 check_is_positive(number::Number) = check_is_positive(string(number))

Checks if the input string is a number that can be parsed as a positive Float64.

Raytracing.string2positive — Function

 string2positive(string::String) :: Float64
 string2positive(number::Number, uint::Bool=false) = 
      string2positive(string(number), uint)

Checks if the input string is a number that can be parsed as a positive Float64 with check_is_positive, and return it as a Float64.

Raytracing.check_is_uint64 — Function

 check_is_uint64(string::String="") :: Bool
 check_is_uint64(number::Number) = check_is_uint64(string(number))

Checks if the input string is a number that can be parsed as a positive Int64.

Raytracing.string2int64 — Function

 string2int64(string::String, uint::Bool=false) :: Union{Int64, UInt64}
 string2int64(number::Number, uint::Bool=false) = 
      string2int64(string(number), uint)

Checks if the input string is a number that can be parsed as a positive Int64 with check_is_uint64, and return it as a Int64 if uint==falseor as a UInt64 if uint==true.

Raytracing.check_is_even_uint64 — Function

 check_is_even_uint64(string::String="") :: Bool
 check_is_even_uint64(number::Number) = check_is_even_uint64(string(number))

Checks if the input string is a number that can be parsed as an even positive Int64.

Raytracing.string2evenint64 — Function

 string2evenint64(string::String, uint::Bool=false) :: Union{Int64, UInt64}
 string2evenint64(number::Number, uint::Bool=false) = string2evenint64(string(number), uint)

Checks if the input string is a number that can be parsed as an even positive Int64 with check_is_even_uint64, and return it as a Int64 if uint==falseor as a UInt64 if uint==true.

Raytracing.check_is_square — Function

 check_is_square(string::String="") :: Bool
 check_is_square(number::Number) = check_is_square(string(number))

Checks if the input string is a number that can be parsed as a squared positive Int64.

Raytracing.string2rootint64 — Function

 string2rootint64(string::String) :: Int64
 string2rootint64(number::Number) = string2rootint64(string(number))

Checks if the input string is a number that can be parsed as a squared positive Int64 with check_is_square, and return the square root as a Int64.

Raytracing.check_is_color — Function

 check_is_color(string::String="") :: Bool

Checks if the input string is a color written in RGB components as "<R, G, B>".

Raytracing.string2color — Function

 string2color(string::String="") :: RGB{Float32}

Checks if the input string is a color written in RGB components as "<R, G, B>" with check_is_color, and return it.

Raytracing.check_is_vector — Function

 check_is_vector(string::String="") :: Bool

Checks if the input string is a vector written in X,Y,Z components as "[X, Y, Z]".

Raytracing.string2vector — Function

 string2vector(string::String="") :: Union{Vec, Nothing}

Checks if the input string is a vector written in X,Y,Z components as "[X, Y, Z]" with check_is_vector, and return Vec(X,Y,Z).

See also: Vec

Raytracing.check_is_declare_float — Function

 check_is_declare_float(string::String="")

Checks if the input string is a declaration of one (or more) floats in the form "NAME:VALUE" with check_is_declare_float. Examples:

    --declare_float=name:1.0
    --declare_float=name1:1.0,name2:2.0
    --declare_float=" name1 : 1.0 , name2: 2.0"

Raytracing.declare_float2dict — Function

 declare_float2dict(string::String) :: Union{Dict{String, Float64}, Nothing}

Checks if the input string is a declaration of one (or more) floats in the form "NAME:VALUE" with check_is_declare_float. Return a Dict{String, Float64} that associates each NAME (as keys) with its Float64 value, or nothing if string=="".

Raytracing.check_is_one_of — Function

check_is_one_of(string::String, vec::Vector{String}) :: Bool

Checks if the input string is inside one of the strings contained in vec.

Raytracing.string2stringoneof — Function

string2stringoneof(string::String, vec::Vector{String}) :: String

Checks if the input string is inside one of the strings contained in vec with check_is_one_of, and return it.

Raytracing.check_is_iterable — Function

 check_is_iterable(string::String, type::Union{Type, Nothing} = nothing) :: Bool
 check_is_iterable(object::T, type::Union{Type, Nothing} = nothing) where T<:Any 
      = check_is_iterable(string(object), type)

Checks if the input string can be parsed in a iterable object, returning true if it is, otherwise false. If specified an input type, check also if all the elements contained in object are of a type T such that T <: type.

Raytracing.string2iterable — Function

 string2iterable(string::String="") :: Vector{String}

Checks if the input string is a vector of variable names written as "[namevar1, namevar2, ...]" with check_is_vec_variables, and return a Vector{String} = ["namevar1", "namevar2", ...].

Raytracing.check_is_vec_variables — Function

 check_is_vec_variables(string::String="") :: Bool

Checks if the input string is a vector of variable names written as "[namevar1, namevar2, ...]".

Raytracing.string2vec_variables — Function

 string2vector(string::String="") :: Union{Vector{String}, Nothing}

Checks if the input string is a vector of variable names written as "[namevar1, namevar2, ...]" with check_is_vec_variables, and return a Vector{String} = ["namevar1", "namevar2", ...].

Raytracing.check_is_function — Function

 check_is_function(string::String="") :: Bool

Checks if the input string is a function name defined in Raytracing.

Raytracing.string2function — Function

 string2function(string::String="") :: Function

Checks if the input string is a function name defined in Raytracing with check_is_function, and return it.

Raytracing.from_CLI_to_vecstring — Function

 from_CLI_to_vecstring(string::String) :: Vector{String}

Parse a string as it would be from the Command Line, and return the Vector{String} that contains all the commands parsed.