Reading and Writing

Raytracing.parse_command_line — Function

parse_command_line(args::Vector{String}) :: 
    (String, String, Float64, Float64)

Interpret the command line when the main is executed, and manage eventual argument errors.

Input

A args::Vector{String}) with length 2, 3 or 4.

Returns

A tuple (infile, outfile, a, γ) containing:

infile::String : first string (required), is the input file name, which must be a PFM format.

outfile::String : second string (required), is the output filename; its format can be PNG or TIFF.

a::Float64 : third argument (optional), is the scale factor for luminosity correction (default 0.18), passed to normalize_image!

γ::Float64 : fourth argument (optional), is the gamma factor for screen correction (default 1.0), passed to γ_correction!

See also : normalize_image!, γ_correction!

Missing docstring.

Missing docstring for parse_demo_settings. Check Documenter's build log for details.

Raytracing.parse_tonemapping_settings — Function

parse_tonemapping_settings(dict::Dict{String, Any}) 
    :: (String, String, Float64, Float64)

Parse a Dict{String, T} where {T} for the tone_mapping function.

Input

A dict::Dict{String, T} where {T}

Returns

A tuple (pfm, png, a, γ) containing:

pfm::String = dict["pfm_infile"] : input pfm filename (required)

png::String = dict["outfile"] : output LDR filename (required)

a::Float64 = dict["alpha"] : scale factor (default = 0.18)

γ::Float64 = dict["gamma"] : gamma factor (default = 1.0)

See also: tone_mapping

Missing docstring.

Missing docstring for parse_demoanimation_settings. Check Documenter's build log for details.

Raytracing.load_image — Function

load_image(path::String) :: HDRimage

Load an image from the specified path in an HDR image format.

See also: HDRimage

Raytracing.ldr2pfm — Function

ldr2pfm(path::String, outfile::String)

Load an image from the specified path, convert it in a pfm file format and save it as outfile. It works through the load_image and the write(::IO, ::HDRimage) function.

See also: HDRimage, load_image, write(::IO, ::HDRimage)