Reading and Writing PFM file

Raytracing.InvalidPfmFileFormat — Type

InvalidPfmFileFormat <: Exception (var::String)

Self-made exception raised by thne following function in case a reading operation of a PFM file fails:

parse_img_size

parse_endianness

read_line

read_float

read(::IO, ::Type{HDRimage})

Raytracing.valid_coordinates — Function

valid_coordinates(img::HDRimage, x::Int, y::Int) :: Bool

Return true if (x, y) are valid coordinates for the 2D matrix of the HDRimage, else return false.

See also: HDRimage

Raytracing.pixel_offset — Function

pixel_offset(img::HDRimage, x::Int, y::Int) :: Int64

Return the index in the 1D array of the specified pixel (x, y) for the given HDRimage. Internally checks also if (x, y) are valid coordinates for img through the valid_coordinates function.

See also: valid_coordinates, HDRimage

Raytracing.get_pixel — Function

get_pixel(img::HDRimage, x::Int, y::Int) :: RBG{Float32}

Return the RBG{Float32} color for the (x, y) pixel in the given HDRimage, obtained through the pixel_offset function. The indexes for a HDRimage pixel matrix with w width and h height follow this sketch:

|  (0,0)    (0,1)    (0,2)   ...   (0,w-1)  |
|  (1,0)    (1,1)    (1,2)   ...   (1,w-1)  |
|   ...      ...      ...    ...     ...    |
| (h-1,0)  (h-1,1)  (h-1,2)  ...  (h-1,w-1) |

See also: pixel_offset, HDRimage

Raytracing.set_pixel — Function

set_pixel(img::HDRimage, x::Int, y::Int, c::RGB{Float32})
set_pixel(img::HDRimage, x::Int, y::Int, c::RGB{T}) where {T}

Set the new RGB color c for the (x, y) pixel in the given HDRimage, acceded through the pixel_offset function. The indexes for a HDRimage pixel matrix with w width and h height follow this sketch:

|  (0,0)    (0,1)    (0,2)   ...   (0,w-1)  |
|  (1,0)    (1,1)    (1,2)   ...   (1,w-1)  |
|   ...      ...      ...    ...     ...    |
| (h-1,0)  (h-1,1)  (h-1,2)  ...  (h-1,w-1) |

If c is of a type RGB{T} where T ≠ Float32, it's called the convert(RGB{Float32}, c) function, which raises an exception if the conversion is not possible.

See also: pixel_offset, HDRimage

Base.write — Method

write(io::IO, img::HDRimage)

Write the given HDRimage image using the given IO stream as a PFM file. The endianness used for writing the file is Little Endian (-1.0).

See also: get_pixel, HDRimage

Raytracing.parse_img_size — Function

parse_img_size(line::String) :: (Int64, Int64)

Return the size (width, height) parsed from a given String, throwing InvalidPfmFileFormat exception if encounters invalid values. It works inside the read(::IO, ::Type{HDRimage}) function.

See also: read(::IO, ::Type{HDRimage}), InvalidPfmFileFormat

Raytracing.parse_endianness — Function

parse_endianness(ess::String) :: Float64

Return the endianness parsed from a given String, throwing InvalidPfmFileFormat exception if encounters an invalid value. It works inside the read(::IO, ::Type{HDRimage}) function.

See also: read(::IO, ::Type{HDRimage}), InvalidPfmFileFormat

Raytracing.read_float — Function

read_float(io::IO, ess::Float64) :: Float32

Return a Float32, readed from the given IO stream io with the (required) endianness ess; ess must be +1.0 (Big Endian) or -1.0 (Little Endian). Controls also if there are enough bits in order to form a Float32, otherwise throw InvalidPfmFileFormat. It works inside the read(::IO, ::Type{HDRimage}) function.

See also: read(::IO, ::Type{HDRimage}), InvalidPfmFileFormat

Raytracing.read_line — Function

read_line(io::IO) :: String

Reads a line from the file whose the given IO object io refers to. Do understand when the file is ended and when a new line begins. Return the readed line as a String. It works inside the read(::IO, ::Type{HDRimage}) function.

See also: read(::IO, ::Type{HDRimage}), InvalidPfmFileFormat

Raytracing.parse_command_line — Function

parse_command_line(args::Vector{String}) :: 
    (String, String, Float64, Float64)

Interpret the command line when the main is executed, and manage eventual argument errors.

Input

A args::Vector{String}) with length 2, 3 or 4.

Returns

A tuple (infile, outfile, a, γ) containing:

infile::String : first string (required), is the input file name, which must be a PFM format.

outfile::String : second string (required), is the output filename; its format can be PNG or TIFF.

a::Float64 : third argument (optional), is the scale factor for luminosity correction (default 0.18), passed to normalize_image!

γ::Float64 : fourth argument (optional), is the gamma factor for screen correction (default 1.0), passed to γ_correction!

See also : normalize_image!, γ_correction!

Base.read — Method

read(io::IO, ::Type{HDRimage}) :: HDRimage

Read a PFM image from a stream object io, and return a HDRimage object containing the image. If an error occurs, raise a InvalidPfmFileFormat exception. Calls internally the following functions:

parse_img_size

parse_endianness

read_line

read_float

See also: parse_img_size, parse_endianness, read_line, read_float, InvalidPfmFileFormat, HDRimage

Raytracing.load_image — Function

load_image(path::Union{String, IO}) :: HDRimage

Load an image from the specified path in an HDR image format.

See also: HDRimage

Raytracing.ldr2pfm — Function

ldr2pfm(path::String, outfile::String)

Load an image from the specified path, convert it in a pfm file format and save it as outfile. It works through the load_image and the write(::IO, ::HDRimage) function.

See also: HDRimage, load_image, write(::IO, ::HDRimage)