The render function

Raytracing.render — Function

render(
      scenefile::String,
      renderer::Renderer = FlatRenderer(),
 	camera_type::Union{String, Nothing} = nothing,
	camera_position::Union{Point, Vec, Nothing} = nothing, 
 	α::Float64 = 0., 
 	width::Int64 = 640, 
 	height::Int64 = 480, 
      a::Float64 = 0.18,
      γ::Float64 = 1.27,
      lum::Union{Number, Nothing} = nothing,
 	pfm_output::String = "scene.pfm", 
    	png_output::String = "scene.png",
      samples_per_pixel::Int64 = 0,
	bool_print::Bool = true,
	bool_savepfm::Bool = true,
      declare_float::Union{Dict{String,Float64}, Nothing} = nothing,
      ONLY_FOR_TESTS::Bool = false,
      )

render(x::(Pair{T1,T2} where {T1,T2})...) = 
     render( parse_render_settings(  Dict( pair for pair in [x...]) )... )

Render the input scenefile with the specified options, and creates the following three files:

the PFM image (scene.pfm is the default name, if none is specified from the command line)

the LDR image (scene.png is the default name, if none is specified from the command line)

the JSON file (which has the same name of the LDR image and .json estention, so scene.json is the default name, if none LDR image name is specified from the command line), that saves some datas about input commands, rendering time etc.

Arguments

The following input arguments refers to the first method presented in the signature; it's obviously very uncomfortable to use that method, so it's recommended to take advantage of the second one, which allows to write the input values in a dictionary like syntax with arbitrary order and comfort. See the documentation of parse_render_settings to learn how to use the keys:

renderer::Renderer = FlatRenderer() : renderer to be used in the rendering, with all the settings already setted (exception made for the world, that will be overridden and created here)

camera_type::String = "per" : set the perspective projection view:

camera_type=="per" -> set PerspectiveCamera (default value)

camera_type=="ort" -> set OrthogonalCamera

camera_position::Union{Point, Vec} = Point(-1.,0.,0.) : set the point of observation in (X,Y,Z) coordinates, as aPointor aVec` input object.

α::Float64 = 0. : angle of rotation IN RADIANTS, relative to the vertical (i.e. z) axis with a right-handed rule convention (clockwise rotation for entering (x,y,z)-axis corresponds to a positive input rotation angle)

width::Int64 = 640 and height::Int64 = 480 : pixel dimensions of the demo image; they must be both even positive integers.

a::Float64 = 0.18 : normalization scale factor for the tone mapping.

γ::Float64 = 1.27 : gamma factor for the tone mapping.

lum::Union{Number, Nothing} = nothing : average luminosity of the image; iIf not specified or equal to 0, it's calculated through avg_lum

pfm_output::String = "demo.pfm" : name of the output pfm file

png_output::String = "demo.png" : name of the output LDR file

samples_per_pixel::Int64 = 0 : number of rays per pixel to be used (antialiasing); it must be a perfect square positive integer (0, 1, 4, 9, ...) and if is set to 0 (default value) is choosen, no anti-aliasing occurs, and only one pixel-centered ray is fired for each pixel.

bool_print::Bool = true : specifies if the WIP messages of the demo function should be printed or not (useful option for render_animation)

bool_savepfm::Bool = true : bool that specifies if the pfm file should be saved or not (useful option for render_animation)

declare_float::Union{Dict{String,Float64}, Nothing} = nothing : an option (for the command line in particularly) to manually override the values of the float variables in the scene file; each overriden variable name (the key) is associated with its float value (i.e. declare_float = Dict("var1"=>0.1, "var2"=>2.5))

ONLY_FOR_TESTS::Bool = false : it's a bool variable conceived only to test the correct behaviour of the renderer for the input arguments; if set to true, no rendering is made!

See also: Point, Vec, Renderer OnOffRenderer, FlatRenderer, PathTracer, PointLightRenderer, parse_render_settings

Raytracing.parse_render_settings — Function

parse_render_settings(dict::Dict{String, T}) where {T}
    :: (
        String, Renderer, String, Vec, Float64, Int64, Int64, String,
        String, Int64, Bool, Bool, Dict{String, Float64}, Bool,
        )

Parse a Dict{String, T} where {T} for the render function.

Input

A dict::Dict{String, T} where {T}

Returns

A tuple (scenefile, renderer, camera_type, camera_position, α, width, height, pfm, png, samples_per_pixel, world_type, bool_print, bool_savepfm, declare_float, ONLY_FOR_TESTS) containing the following variables (the corresponding keys are also showed):

scenefile::String = dict["scenefile"] : name of the scene file to be rendered; it must be written with the correct syntax, see the tutorial_basic_sintax.txt and the demo_world_B.txt files.

renderer::Renderer = haskey(dict, "renderer") ? dict["renderer"] : dict["%COMMAND%"] : it's the renderer to be used (with a default empy World that will be populated in the demo function); the possible keys are two, and must be used differently:

the dict["renderer"] must contain the renderer itself (i.e. a Renderer object); if this key exists, it has the priority on the latter key

the dict["%COMMAND%"] must contain a string that identifies the type of renderer to be used, i.e. one of the following strings:

"dict["%COMMAND%"]=>onoff" -> OnOffRenderer algorithm

"dict["%COMMAND%"]=>"flat" -> FlatRenderer algorithm (default value)

"dict["%COMMAND%"]=>"pathtracing" -> PathTracer algorithm

"dict["%COMMAND%"]=>"pointlight" -> PointLightRenderer algorithm

Moreover, in this second case, you can specify the options for the correspoding renderer through another dictionary associated with the kkey of the renderer name; that options will be parsed thanks to the corresponding functions. We shows in the next lines the key-value syntax described:

"onoff"=>Dict{String, T} where {T} : parsed with parse_onoff_settings

"flat"=>Dict{String, T} where {T} : parsed with parse_flat_settings

"pathtracer"=>Dict{String, T} where {T} : parsed with parse_pathtracer_settings

"pointlight"=>Dict{String, T} where {T} : parsed with parse_pointlight_settings

camera_type::String = dict["camera_type"] : set the perspective projection view; it must be one of the following values, and this is checked with the string2stringoneof function:

"per" -> set PerspectiveCamera (default value)

"ort" -> set OrthogonalCamera

camera_position::String = typeof(dict["camera_position"]) ∈ [Vec, Point] ? dict["camera_position"] : string2vector(dict["camera_position"]) : "[X, Y, Z]" coordinates of the choosen observation point of view; it can be specified in two ways:

if typeof(dict["camera_position"]) is a Vec or a Point, it's passed as-is

else, it must be a String written in the form "[X, Y, Z]" , and it's parsed through string2vectorto aVec` object

α::String = dict["alpha"] : choosen angle of rotation IN RADIANTS respect to vertical (i.e. z) axis with a right-handed rule convention (clockwise rotation for entering (x,y,z)-axis corresponds to a positive input rotation angle)

width::Int64 = string2evenint64(dict["width"]) : number of pixels on the horizontal axis to be rendered; it's converted through string2evenint64 to a even positive integer.

height::Int64 = string2evenint64(dict["height"]) : number of pixels on the vertical axis to be rendered; it's converted through string2evenint64 to a even positive integer.

a::Float64 = string2positive(dict["normalization"]) : scale factor (default = 0.18); it's converted through string2positive to a positive floating point number.

γ::Float64 = string2positive(dict["gamma"]) : gamma factor (default = 1.0); it's converted through string2positive to a positive floating point number.

lum::Union{Number, Nothing} = string2positive(dict["avg_lum"]) : average luminosity of the image; it's converted through string2positive to a positive floating point number. If not specified or equal to 0, it's calculated through avg_lum

pfm::String = dict["set_pfm_name"] : output pfm filename (default "scene.pfm")

png::String = dict["setpngname"]: output LDR filename (default"scene.png"`)

samples_per_pixel::Int64 = dict["samples_per_pixel"] : number of ray to be generated for each pixel, implementing the anti-aliasing algorithm; it must be a perfect integer square (0,1,4,9,...) and this is checked with the string2rootint64 function; if 0 (default value) is choosen, no anti-aliasing occurs, and only one pixel-centered ray is fired for each pixel.

bool_print::Bool = dict["bool_print"] : if true (default value), WIP message of demo function are printed (otherwise no; it's useful for the demo_animation function)

bool_savepfm::Bool = dict["bool_savepfm"] : if true (default value), demo function saves the pfm file to disk (otherwise no; it's useful for the demo_animation function)

declare_float::Union{Dict{String, Float64}, Nothing} = declare_float2dict(dict["declare_float"]) : an option for the command line to manually override the values of the float variables in the scene file. The input dict["declare_float"] must be a String written such as "var1:0.1, var2 : 2.5"; such a string is parsed through the declare_float2dict in a Dict{String, Float64} where each overriden variable name (the key) is associated with its float value (declare_float=>Dict("var1"=>0.1, "var2"=>2.5))

ONLY_FOR_TESTS::Bool = dict["ONLY_FOR_TESTS"] : it's a bool variable conceived only to test the correct behaviour of the renderer for the input arguments; if set to true, no rendering is made!

See also: render, Renderer, Vec, string2evenint64, string2stringoneof, string2positive, string2vector, string2rootint64, declare_float2dict

Raytracing.parse_onoff_settings — Function

parse_onoff_settings(dict::Dict{String, T}) where {T}
    :: (World, RGB{Float32}, RGB{Float32})

Parse a Dict{String, T} where {T} for the initialisation of a OnOffRenderer.

Input

A dict::Dict{String, T} where {T}

Returns

A tuple (World(), background_color, color) containing the following variables (the corresponding keys are also showed):

World():: World : is the default constructor of the World class, which creates an empty world; it will be populated in the function that will use the renderer.

background_color::RGB{Float32} = string2color(dict["background_color"]) : set the color returned by a light ray which does not hit any object in the scene; the default value is BLACK, i.e. RGB{Float32}(0.0, 0.0, 0.0). The input color value dict["background_color"] must be a String written in RGB components as "< R , G , B >", and it's parsed with the string2color function.

color::RGB{Float32} = string2color(dict["color"]) : set the color returned by a light ray which does hit any object in the scene; the default value is WHITE, i.e. RGB{Float32}(1.0, 1.0, 1.0). The input color value dict["color"] must be a String written in RGB components as "< R , G , B >", and it's parsed with the string2color function.

See also: Renderer, OnOffRenderer, World, string2color

Raytracing.parse_flat_settings — Function

parse_flat_settings(dict::Dict{String, T}) where {T}
    :: (World, RGB{Float32})

Parse a Dict{String, T} where {T} for the initialisation of a FlatRenderer.

Input

A dict::Dict{String, T} where {T}

Returns

A tuple (World(), background_color) containing the following variables (the corresponding keys are also showed):

World():: World : is the default constructor of the World class, which creates an empty world; it will be populated in the function that will use the renderer.

background_color::RGB{Float32} = string2color(dict["background_color"]) : set the color returned by a light ray which does not hit any object in the scene; the default value is BLACK, i.e. RGB{Float32}(0.0, 0.0, 0.0). The input color value dict["background_color"] must be a String written in RGB components as "< R , G , B >", and it's parsed with the string2color function.

See also: Renderer, FlatRenderer, World, string2color

Raytracing.parse_pathtracer_settings — Function

parse_pathtracer_settings(dict::Dict{String, T}) where {T}
    :: (World(), RGB{Float32}, PCG, Int64, Int64, Int64)

Parse a Dict{String, T} where {T} for the initialisation of a PathTracer.

Input

A dict::Dict{String, T} where {T}

Returns

A tuple (World(), background_color, PCG(init_state, init_seq), num_of_rays, max_depth, russian_roulette_limit, ONLY_FOR_TESTS) containing the following variables (the corresponding keys are also showed):

World():: World : is the default constructor of the World class, which creates an empty world; it will be populated in the function that will use the renderer.

background_color::RGB{Float32} = string2color(dict["background_color"]) : set the color returned by a light ray which does not hit any object in the scene; the default value is BLACK, i.e. RGB{Float32}(0.0, 0.0, 0.0). The input color value dict["background_color"] must be a String written in RGB components as "< R , G , B >", and it's parsed with the string2color function.

PCG(init_state, init_seq)::PCG : a mutable struct of the Permuted Congruential Generator (PCG), which is a uniform pseudo-random number generator; you can pass as input two usigned integer that initialize the generator:

init_state::UInt64 = string2int64(dict["init_state"], true) : set the initial state of the PCG; the input value dict["init_state"] is parsed thanks to the string2int64 function.

init_seq::UInt64 = string2int64(dict["init_seq"], true) : set the initial sequence of the PCG; the input value dict["init_seq"] is parsed thanks to the string2int64 function.

num_of_rays::Int64 = string2int64(dict["num_of_rays"]) : set the number of secondary rays that will be fired from each surface point hitted by a light ray; the input value dict["num_of_rays"] is parsed thanks to the string2int64 function.

max_depth::Int64 = string2int64(dict["max_depth"]) : set the maximum depth number that the secondary rays are allowed to have; if that value is exceeded, no more secondary rays are generated in the hitten points, and the returned color is BLACK; the input value dict["max_depth"] is parsed thanks to the string2int64 function.

russian_roulette_limit::Int64 = string2int64(dict["russian_roulette_limit"]) : set the depth over which a secondary ray is created with the russian roulette algorithm; the input value dict["russian_roulette_limit"] is parsed thanks to the string2int64 function.

See also: Renderer, PathTracer, World, PCG, string2color, string2int64

Raytracing.parse_pointlight_settings — Function

parse_pointlight_settings(dict::Dict{String, T}) where {T}
    :: (World, RGB{Float32}, RGB{Float32})

Parse a Dict{String, T} where {T} for the initialisation of a PointLightRenderer.

Input

A dict::Dict{String, T} where {T}

Returns

A tuple (World(), background_color, color, ONLY_FOR_TESTS) containing the following variables (the corresponding keys are also showed):

World():: World : is the default constructor of the World class, which creates an empty world; it will be populated in the function that will use the renderer.

background_color::RGB{Float32} = string2color(dict["background_color"]) : set the color returned by a light ray which does not hit any object in the scene; the default value is BLACK, i.e. RGB{Float32}(0.0, 0.0, 0.0). The input color value dict["background_color"] must be a String written in RGB components as "< R , G , B >", and it's parsed with the string2color function.

ambient_color::RGB{Float32} = string2color(dict["ambient_color"]) : set the minimum color returned by a light ray which hits an object on a point, indipendently that is or is not directy visible from any of the point-light sources in the scene; the default value is BLACK, i.e. RGB{Float32}(0.0, 0.0, 0.0). The input color value dict["ambient_color"] must be a String written in RGB components as "< R , G , B >", and it's parsed with the string2color function.

dark_parameter::Float64 = string2positive(dict["dark_parameter"]) : float that defines the retuned percentage of the hit point color if it is not visible from any of the point-light source of the image; a non-zero value allows to see also the not-directly illuminated parts of the image. The default value is 0.05.

See also: Renderer, PointLightRenderer, World, string2color

The render_animation function

Raytracing.render_animation — Function

render_animation( 
		func::Function,
           vec_variables::Vector{String},
           iterable::Any,
           scenefile::String,
           renderer_model::Renderer = FlatRenderer(),
           camera_type::Union{String, Nothing} = nothing,
           camera_position::Union{Point, Vec, Nothing} = nothing, 
           α::Float64 = 0., 
           width::Int64 = 640, 
           height::Int64 = 480, 
           a::Float64 = 0.18,
           γ::Float64 = 1.27,
           lum::Union{Number, Nothing} = nothing,
           anim_output::String = "scene_animation.mp4", 
           samples_per_pixel::Int64 = 0,
           bool_print::Bool = true,
           declare_float::Union{Dict{String,Float64}, Nothing} = nothing,
           ONLY_FOR_TESTS::Bool = false,     
	)

render_animation(x::(Pair{T1,T2} where {T1,T2})...) = 
	render_animation( parse_render_animation_settings(  Dict( pair for pair in [x...]) )... )

Render the input scenefile as an animation with the specified options, and creates the following three files:

the animation (scene_animation.mp4 is the default name, if none is specified from the command line)

the JSON file (which has the same name of the animation and .json estention, so scene_animation.json is the default name), that saves some datas about input commands, rendering time etc.

This function works following this steps:

creates an hidden directory, called ".wip_animation"; if it already exists, it will be destroyed and recreated.

inside ".wpi_animation", creates the png images of the rendered image (using the render function with the specified projection, renderer and image dims); each image correspons to a frame of the future animation

through the ffmpeg software, the 360 png images are converted into the animation mp4 file, and saved in the main directory

the ".wpi_animation" directory and all the png images inside it are destroyed

Arguments

The following input arguments refers to the first method presented in the signature; it's obviously very uncomfortable to use that method, so it's recommended to take advantage of the second one, which allows to write the input values in a dictionary like syntax with arbitrary order and comfort. See the documentation of parse_render_animation_settings to learn how to use the keys:

func::Function : function that takes as input the frame number and returns the values of the variables for that frame; it must have a method that takes only one input number, and must return a tuple of length equals to the vec_variables one.

vec_variables::Vector{String} : vector that contains the variable names (DEFINED IN SCENEFILE) that will be overridden from frame to frame; its length must equals the tuple length returned by func.

iterable::Any : an iterable object, that defines the frame numbers and the total number of frames; its values will be given in input to func.

renderer::Renderer = FlatRenderer() : renderer to be used in the rendering, with all the settings already setted (exception made for the world, that will be overridden and created here)

camera_type::String = "per" : set the perspective projection view:

camera_type=="per" -> set PerspectiveCamera (default value)

camera_type=="ort" -> set OrthogonalCamera

camera_position::Union{Point, Vec} = Point(-1.,0.,0.) : set the point of observation in (X,Y,Z) coordinates, as aPointor aVec` input object.

α::Float64 = 0. : angle of rotation IN RADIANTS, relative to the vertical (i.e. z) axis with a right-handed rule convention (clockwise rotation for entering (x,y,z)-axis corresponds to a positive input rotation angle)

width::Int64 = 640 and height::Int64 = 480 : pixel dimensions of the demo image; they must be both even positive integers.

a::Float64 = 0.18 : normalization scale factor for the tone mapping.

γ::Float64 = 1.27 : gamma factor for the tone mapping.

lum::Union{Number, Nothing} = nothing : average luminosity of the image; iIf not specified or equal to 0, it's calculated through avg_lum

anim_output::String = "scene_animation.mp4" : name of the output animation file.

samples_per_pixel::Int64 = 0 : number of rays per pixel to be used (antialiasing); it must be a perfect square positive integer (0, 1, 4, 9, ...) and if is set to 0 (default value) is choosen, no anti-aliasing occurs, and only one pixel-centered ray is fired for each pixel.

bool_print::Bool = true : specifies if the WIP messages of the demo function should be printed or not.

declare_float::Union{Dict{String,Float64}, Nothing} = nothing : an option (for the command line in particularly) to manually override the values of the float variables in the scene file; each overriden variable name (the key) is associated with its float value (i.e. declare_float = Dict("var1"=>0.1, "var2"=>2.5))

ONLY_FOR_TESTS::Bool = false : it's a bool variable conceived only to test the correct behaviour of the renderer for the input arguments; if set to true, no rendering is made!

See also: Point, Vec, Renderer OnOffRenderer, FlatRenderer, PathTracer, PointLightRenderer, parse_render_animation_settings, render

Raytracing.parse_render_animation_settings — Function

parse_render_animation_settings(dict::Dict{String, T}) where {T}
    :: (
        Function, Vector{String}, Any,
        String, Renderer, String, Vec, Float64, Int64, Int64, String,
        Int64, Bool, Dict{String, Float64}, Bool,
        )

Parse a Dict{String, T} where {T} for the render_animation function.

Input

A dict::Dict{String, T} where {T}

Returns

A tuple (func, vec_variables, iterable, scenefile, renderer, camera_type, camera_position, α, width, height, anim, samples_per_pixel, bool_print, declare_float, ONLY_FOR_TESTS) containing the following variables (the corresponding keys are also showed):

scenefile::String = dict["scenefile"] : name of the scene file to be rendered; it must be written with the correct syntax, see the tutorial_basic_sintax.txt and the demo_world_B.txt files.

renderer::Renderer = haskey(dict, "renderer") ? dict["renderer"] : dict["%COMMAND%"] : it's the renderer to be used (with a default empy World that will be populated in the demo function); the possible keys are two, and must be used differently:

the dict["renderer"] must contain the renderer itself (i.e. a Renderer object); if this key exists, it has the priority on the latter key

the dict["%COMMAND%"] must contain a string that identifies the type of renderer to be used, i.e. one of the following strings:

"dict["%COMMAND%"]=>onoff" -> OnOffRenderer algorithm

"dict["%COMMAND%"]=>"flat" -> FlatRenderer algorithm (default value)

"dict["%COMMAND%"]=>"pathtracing" -> PathTracer algorithm

"dict["%COMMAND%"]=>"pointlight" -> PointLightRenderer algorithm

Moreover, in this second case, you can specify the options for the correspoding renderer through another dictionary associated with the kkey of the renderer name; that options will be parsed thanks to the corresponding functions. We shows in the next lines the key-value syntax described:

"onoff"=>Dict{String, T} where {T} : parsed with parse_onoff_settings

"flat"=>Dict{String, T} where {T} : parsed with parse_flat_settings

"pathtracer"=>Dict{String, T} where {T} : parsed with parse_pathtracer_settings

"pointlight"=>Dict{String, T} where {T} : parsed with parse_pointlight_settings

camera_type::String = dict["camera_type"] : set the perspective projection view; it must be one of the following values, and this is checked with the string2stringoneof function:

"per" -> set PerspectiveCamera (default value)

"ort" -> set OrthogonalCamera

camera_position::String = typeof(dict["camera_position"]) ∈ [Vec, Point] ? dict["camera_position"] : string2vector(dict["camera_position"]) : "[X, Y, Z]" coordinates of the choosen observation point of view; it can be specified in two ways:

if typeof(dict["camera_position"]) is a Vec or a Point, it's passed as-is

else, it must be a String written in the form "[X, Y, Z]" , and it's parsed through string2vectorto aVec` object

α::String = dict["alpha"] : choosen angle of rotation IN RADIANTS respect to vertical (i.e. z) axis with a right-handed rule convention (clockwise rotation for entering (x,y,z)-axis corresponds to a positive input rotation angle)

width::Int64 = string2evenint64(dict["width"]) : number of pixels on the horizontal axis to be rendered; it's converted through string2evenint64 to a even positive integer.

height::Int64 = string2evenint64(dict["height"]) : number of pixels on the vertical axis to be rendered; it's converted through string2evenint64 to a even positive integer.

a::Float64 = string2positive(dict["normalization"]) : scale factor (default = 0.18); it's converted through string2positive to a positive floating point number.

γ::Float64 = string2positive(dict["gamma"]) : gamma factor (default = 1.0); it's converted through string2positive to a positive floating point number.

lum::Union{Number, Nothing} = string2positive(dict["avg_lum"]) : average luminosity of the image; it's converted through string2positive to a positive floating point number. If not specified or equal to 0, it's calculated through avg_lum

anim::String = dict["set_anim_name"] : output animation name (default "animation.mp4")

samples_per_pixel::Int64 = dict["samples_per_pixel"] : number of ray to be generated for each pixel, implementing the anti-aliasing algorithm; it must be a perfect integer square (0,1,4,9,...) and this is checked with the string2rootint64 function; if 0 (default value) is choosen, no anti-aliasing occurs, and only one pixel-centered ray is fired for each pixel.

bool_print::Bool = dict["bool_print"] : if true (default value), WIP message of demo function are printed (otherwise no; it's useful for the demo_animation function)

declare_float::Union{Dict{String, Float64}, Nothing} = declare_float2dict(dict["declare_float"]) : an option for the command line to manually override the values of the float variables in the scene file. The input dict["declare_float"] must be a String written such as "var1:0.1, var2 : 2.5"; such a string is parsed through the declare_float2dict in a Dict{String, Float64} where each overriden variable name (the key) is associated with its float value (declare_float=>Dict("var1"=>0.1, "var2"=>2.5))

ONLY_FOR_TESTS::Bool = dict["ONLY_FOR_TESTS"] : it's a bool variable conceived only to test the correct behaviour of the renderer for the input arguments; if set to true, no rendering is made!

See also: render_animation, Renderer, Vec, string2evenint64, string2stringoneof, string2positive, string2vector, string2rootint64, declare_float2dict, string2iterable, string2vec_variables