Renderers

Raytracing.Renderer — Type

abstract type Renderer <: Function end

An abstract class implementing a solver of the rendering equation. The concrete sub-types of this abstract class are:

OnOffRenderer

FlatRenderer

PathTracer

PointLightRenderer

Raytracing.OnOffRenderer — Type

OnOffRenderer <: Renderer(
    world::World = World()
    background_color::RGB{Float32} = RGB{Float32}(0.0, 0.0, 0.0)
    color::RGB{Float32} = RGB{Float32}(1.0, 1.0, 1.0)
)

A on/off renderer. If the ray intersecty anyone of the shape inside the given world, the returned color will be color; otherwise, if no shape is intersected, the returned color will be background_color.

This renderer is mostly useful for debugging purposes, as it is really fast, but it produces boring images.

See also: Renderer, World

Raytracing.FlatRenderer — Type

FlatRenderer <: Renderer(
    world::World = World()
    background_color::RGB{Float32} = RGB{Float32}(0.0, 0.0, 0.0)
)

A «flat» renderer. This renderer estimates the solution of the rendering equation by neglecting any contribution of the light. It just uses the pigment of each surface to determine how to compute the final radiance.

See also: Renderer, World

Raytracing.PathTracer — Type

PathTracer <: Renderer(
        world::World, 
        background_color::RGB{Float32} = RGB{Float32}(0.0, 0.0, 0.0),
        pcg::PCG = PCG(),
        num_of_rays::Int64 = 10,
        max_depth::Int64 = 2,
        russian_roulette_limit::Int64 = 3
    )

A simple path-tracing renderer.

The algorithm implemented here allows the caller to tune number of rays thrown at each iteration, as well as the maximum depth. It implements Russian roulette, so in principle it will take a finite time to complete the calculation even if you set max_depth to Inf.

Arguments

world::World : the world to be rendered

background_color::RGB{Float32} : default background color if the Ray doesn-t hit anything

pcg::PCG : PCG random number generator for evaluating integrals

num_of_rays::Int64 : number of Rays generated for each integral evaluation

max_depth::Int64 : maximal number recursive integrations; if a ray intersecting a surface has depth>max_depth, the returned color is RGB{Float32}(0,0,0)

russian_roulette_limit::Int64: depth at whitch the Russian Roulette algorithm begins

See also: Renderer, Ray, World, PCG

Raytracing.PointLightRenderer — Type

PointLightRenderer <: Renderer (
    world::World,
    background_color::RGB{Float32} = RGB{Float32}(0., 0., 0.),
    ambient_color::RGB{Float32} = RGB{Float32}(0.1, 0.1, 0.1)
    dark_parameter::Float64 = 0.05
)

A simple point-light tracing renderer.

Arguments

world::World : the world to be rendered

background_color::RGB{Float32} : default background color if the Ray doesn-t hit anything

ambient_color::RGB{Float32} : default ambient color

dark_parameter::Float64 : float that defines the retuned percentage of the hit point color if it is not visible from any of the point-light source of the image; a non-zero value allows to see also the not-directly illuminated parts of the image.

See also: Renderer, World

Raytracing.is_point_visible — Function

is_point_visible(
        world::World, 
        point::Point, 
        observer_pos::Point
        ) :: Bool

Return true if the straight line connecting observer_pos to point do not intersect any of the shapes of world between the two points, otherwise return false.

See also: World, Point

Raytracing.quick_ray_intersection — Function

quick_ray_intersection(shape::Shape, ray::Ray) :: Bool

Quickly determine whether the ray hits the shape or not.

See also: Shape, Ray

quick_ray_intersection(sphere::Sphere, ray::Ray) :: Bool

Quickly checks if the ray intersects the sphere or not.

See also: Sphere, Ray

quick_ray_intersection(plane::Plane, ray::Ray) :: Bool

Quickly checks if the ray intersects the plane or not.

See also: Plane, Ray