Avaiable shapes

Raytracing.BRDF

Raytracing.Camera

Raytracing.CheckeredPigment

Raytracing.DiffuseBRDF

Raytracing.FlatRenderer

Raytracing.HDRimage

Raytracing.HitRecord

Raytracing.ImagePigment

Raytracing.ImageTracer

Raytracing.Material

Raytracing.Normal

Raytracing.OnOffRenderer

Raytracing.OrthogonalCamera

Raytracing.PCG

Raytracing.Parameters

Raytracing.PathTracer

Raytracing.PerspectiveCamera

Raytracing.Pigment

Raytracing.Plane

Raytracing.Point

Raytracing.PointLight

Raytracing.PointLightRenderer

Raytracing.Ray

Raytracing.Renderer

Raytracing.Shape

Raytracing.SpecularBRDF

Raytracing.Sphere

Raytracing.Transformation

Raytracing.Triangle

Raytracing.UniformPigment

Raytracing.Vec

Raytracing.Vec2d

Raytracing.World

Base.read

Base.write

Raytracing.add_light!

Raytracing.add_shape!

Raytracing.at

Raytracing.clamp

Raytracing.clamp_image!

Raytracing.create_onb_from_z

Raytracing.demo

Raytracing.demo_animation

Raytracing.evaluate

Raytracing.fire_all_rays!

Raytracing.fire_ray

Raytracing.first_world

Raytracing.get_matrix

Raytracing.get_pixel

Raytracing.inverse

Raytracing.is_consistent

Raytracing.is_point_visible

Raytracing.ldr2pfm

Raytracing.load_image

Raytracing.lum_max

Raytracing.luminosity

Raytracing.normalize_image!

Raytracing.parse_demo_settings

Raytracing.parse_demoanimation_settings

Raytracing.parse_endianness

Raytracing.parse_img_size

Raytracing.parse_tonemapping_settings

Raytracing.pixel_offset

Raytracing.plane_normal

Raytracing.plane_point_to_uv

Raytracing.quick_ray_intersection

Raytracing.random

Raytracing.ray_intersection

Raytracing.ray_intersection

Raytracing.ray_intersection

Raytracing.read_float

Raytracing.read_line

Raytracing.rotation_x

Raytracing.rotation_y

Raytracing.rotation_z

Raytracing.scaling

Raytracing.scatter_ray

Raytracing.second_world

Raytracing.select_world

Raytracing.set_pixel

Raytracing.sphere_normal

Raytracing.sphere_point_to_uv

Raytracing.third_world

Raytracing.to_RGB

Raytracing.tone_mapping

Raytracing.translation

Raytracing.triangle_barycenter

Raytracing.triangle_normal

Raytracing.triangle_point_to_uv

Raytracing.valid_coordinates

Raytracing.γ_correction!

Sphere

Raytracing.Sphere — Type

Sphere <: Shape(
    T::Transformation = Transformation(),
    Material::Material = Material()
)

A 3D unit sphere, i.e. centered on the origin of the axes and with radius 1.0.

Arguments

T::Transformation : transformation associated to the sphere.

Material::Material : material that constitutes the sphere.

See also: Shape, Transformation, Material

Raytracing.sphere_point_to_uv — Function

sphere_point_to_uv(point::Point) :: Vec2d

Convert a 3D point $P = (P_x, P_y, P_z)$ on the surface of the unit sphere into a 2D Vec2d using the following spherical coordinates:

\[u = \frac{\phi}{2\pi} = \frac{\arctan (P_y / P_x)}{2\pi}, \quad v = \frac{\theta}{\pi} = \frac{\arccos (P_z)}{\pi}\]

See also: Point, Vec2d, Sphere

Raytracing.sphere_normal — Function

sphere_normal(point::Point, ray_dir::Vec) :: Normal

Compute the Normal of a unit sphere.

The normal is computed for the given Point $point = (P_x, P_y, P_z)$ (with $\sqrt{P_x^2 + P_y^2 + P_z^2}=1$) on the surface of the sphere, and it is chosen so that it is always in the opposite direction with respect to the given Vec ray_dir.

See also: Point, Ray, Normal, Sphere

Raytracing.ray_intersection — Method

ray_intersection(sphere::Sphere, ray::Ray) :: Union{HitRecord, Nothing}

Check if the ray intersects the sphere. Return a HitRecord, or nothing if no intersection is found.

See also: Ray, Sphere, HitRecord

Plane

Raytracing.Plane — Type

Plane <: Shape(
    T::Transformation = Transformation(),
    Material::Material = Material()
)

A 3D unit plane, i.e. the x-y plane (set of 3D points with z=0).

Arguments

T::Transformation : transformation associated to the plane.

Material::Material : material that constitutes the plane.

See also: Shape, Transformation, Material

Raytracing.plane_point_to_uv — Function

plane_point_to_uv(point::Point) :: Vec2d

Convert a 3D point $P = (P_x, P_y, P_z)$ on the surface of the unit plane into a 2D Vec2d using the following periodical coordinates:

\[u = P_x - \lfloor P_x \rfloor, \quad v = P_y - \lfloor P_y \rfloor,\]

where $\lfloor \cdot \rfloor$ indicates the rounding down approximation, in order to guarantee that $u, v \in [0, 1)$.

See also: Point, Vec2d, Plane

Raytracing.plane_normal — Function

plane_normal(point::Point, ray_dir::Vec) :: Normal

Compute the Normal of a unit plane.

The normal is computed for the given Point $point = (P_x, P_y, 0)$ on the surface of the plane, and it is chosen so that it is always in the opposite direction with respect to the given Vec ray_dir.

See also: Point, Ray, Normal, Plane

Raytracing.ray_intersection — Method

ray_intersection(plane::Plane, ray::Ray) :: Union{HitRecord, Nothing}

Check if the ray intersects the plane. Return a HitRecord, or nothing if no intersection is found.

See also: Ray, Plane, HitRecord

Triangle

Raytracing.Triangle — Type

Triangle <: Shape(
    vertexes::SVector{3, Point} = 
        SVector{3, Point}(Point(√3/2, 0, 0), Point(0, 0.5, 0), Point(0, -0.5, 0)),
    Material::Material = Material()
)

A 3D triangle.

Arguments

vertexes::SVector{3, Point} : points associated to the triangle.

Material::Material : material that constitutes the triangle.

See also: Shape, Material

Raytracing.triangle_point_to_uv — Function

triangle_point_to_uv(triangle::Triangle, point::Point) :: Vec2d

Return the barycentic coordinates of the given point for the input triangle.

If the triangle is made of the vertexes $(A,B,C)$ (memorized in this order), then the point $P$ has coordinates $(u,v) = (\beta, \gamma)$ such that:

\[ P(\beta, \gamma) = A + \beta \,(B - A) + \gamma \,(C-A)\]

The analitic resolution of this linear system is:

\[\begin{aligned} &\beta = \frac{ (P_x - A_x)(C_y - A_y) - (P_y - A_y)(C_x - A_x) }{ (B_x - A_x)(C_y - A_y) - (B_y - A_y)(C_x - A_x) } \\ &\gamma = \frac{ (P_x - A_x)(B_y - A_y) - (P_y - A_y)(B_x - A_x) }{ (C_x - A_x)(B_y - A_y) - (C_y - A_y)(B_x - A_x) } \end{aligned}\]

NOTE: this function do not check if $P$ is on the plane defined by $(A,B,C)$, neither if $P$ is inside the triangle made of them!

See also: Triangle, Vec2d, Point

Raytracing.triangle_barycenter — Function

triangle_barycenter(triangle::Triangle) :: Point

Return the barycenter of the given triangle.

For a triangle with vertexes $(A, B, C)$, the barycenter is $M$:

\[ M = \frac{A + B + C}{3}\]

See also: Triangle, Point

Raytracing.triangle_normal — Function

triangle_normal(triangle::Triangle, ray_dir::Vec) :: Normal

Compute the Normal of a given triangle.

The normal for a triangle with vertexes $(A, B, C)$ is computed as follows:

\[ n = \pm (B-A) \times (C-A)\]

where the sign is chosen so that it is always in the opposite direction with respect to the given ray_dir.

See also: Point, Ray, Normal, Triangle

Raytracing.ray_intersection — Method

ray_intersection(triangle::Triangle, ray::Ray) :: Union{HitRecord, Nothing}

Check if the ray intersects the triangle. Return a HitRecord, or nothing if no intersection is found.

For a triangle with vertexes $(A, B, C)$ and a ray defined with the simple equation $r(t) = O + t \, \vec{d}$, the coordinates $(u,v) = (\beta, \gamma)$ and the t value of intersection are obtained solving this linear system:

\[ \begin{bmatrix} B_x-A_x & C_x-A_x & -d_x \\ B_y-A_y & C_y-A_y & -d_y \\ B_z-A_z & C_z-A_z & -d_z \end{bmatrix} \begin{bmatrix} u \\ v \\ t \end{bmatrix} = \begin{bmatrix} O_x - A_x\\ O_z - A_z\\ O_z - A_z \end{bmatrix}\]

See also: Ray, Triangle, HitRecord