Tone Mapping

Raytracing.luminosity — Function

luminosity(c::RGB{T}) :: Float64

Return the best average luminosity of a color $c = (R_i, G_i, B_i)$, through the Shirley & Morley proposal:

\[l_i = \frac{\max\{R_i, G_i, B_i \} + \min\{R_i, G_i, B_i \}}{2}\]

Raytracing.lum_max — Function

lum_max(img::HDRimage) :: Float64

Return the maximum luminosity of the given img according to the luminosity function of a color.

See also: HDRimage, luminosity

Missing docstring.

Missing docstring for Raytracing.avr_lum. Check Documenter's build log for details.

Raytracing.clamp — Function

clamp(x::Number) :: Float64

Clamping function:

\[x \rightarrow x/(x+1)\]

Raytracing.normalize_image! — Function

normalize_image!(  
        img::HDRimage, 
        a::Float64 = 0.18,
        lum::Union{Number, Nothing} = nothing, 
        δ::Number = 1e-10
        )

Normalize the img colors through the following formula:

\[\forall \; \mathrm{colors} \, c =(R_i, G_i, B_i) \; \mathrm{of \, the \, image} \\\\ X_i \rightarrow \frac{a}{\langle l \rangle} X_i \quad , \quad \forall X_i = R_i, G_i, B_i \]

where $\langle l \rangle$ is the average luminosity returned by the avg_lum function.

See also: HDRimage, avg_lum

Raytracing.clamp_image! — Function

clamp_image!(img::HDRimage)

Adjust the color levels of the brightest pixels in the img through the clamp function.

See also: HDRimage, clamp

Raytracing.γ_correction! — Function

γ_correction!(img::HDRimage, γ::Float64=1.0, k::Float64=1.0)

Corrects the image using the γ factor, assuming a potential dependence between the input and output signals of a monitor/screen.

As third optional argument, you can pass the maximum value 'k' of the range you want the RGB colors may have. The default value is 'k=1.0', so the range RGB colors can span is '[0.0, 1.0]'

See also: HDRimage

Raytracing.tone_mapping — Function

Tone-map the given input pfm file infile with luminosity normalisation a and gamma factor γ. Return a file with the input outfile name outfile and of the specified LDR format, if possible.

In order to do the tone-mapping, this function relies on the following three function, called in the presented order:

normalize_image! : normalize the image colors

clamp_image! : adjust the color levels of the brightest pixels

γ_correction! : corrects the image using the γ factor

See also: normalize_image!, clamp_image!, γ_correction!

Raytracing.parse_tonemapping_settings — Function

parse_tonemapping_settings(dict::Dict{String, Any}) 
    :: (String, String, Float64, Float64)

Parse a Dict{String, T} where {T} for the tone_mapping function.

Input

A dict::Dict{String, T} where {T}

Returns

A tuple (pfm, png, a, γ) containing:

pfm::String = dict["infile"] : input pfm filename (required)

png::String = dict["outfile"] : output LDR filename (required)

a::Float64 = dict["alpha"] : scale factor (default = 0.18)

γ::Float64 = dict["gamma"] : gamma factor (default = 1.0)

See also: tone_mapping