Transformations

Raytracing.rotation_x — Function

rotation_x(ϑ::Float64) :: Transformation

Encoding a rotation around the x-axis of an angle ϑ in radiant.

The positive sign is given by the right-hand rule, therefore clockwise rotation for entering x-axis corresponds to a ϑ>0 rotation angle.

See also: Transformation

Raytracing.rotation_y — Function

rotation_y(ϑ::Float64) :: Transformation

Encoding a rotation around the y-axis of an angle ϑ in radiant.

The positive sign is given by the right-hand rule, therefore clockwise rotation for entering y-axis corresponds to a ϑ>0 rotation angle.

See also: Transformation

Raytracing.rotation_z — Function

rotation_z(ϑ::Float64) :: Transformation

Encoding a rotation around the z-axis of an angle ϑ in radiant.

The positive sign is given by the right-hand rule, therefore clockwise rotation for entering z-axis corresponds to a ϑ>0 rotation angle.

See also: Transformation

Raytracing.scaling — Function

scaling(v::Vec) :: Transformation
scaling(x::Float64, y::Float64, z::Float64) = scaling(Vec(x,y,z))

Encoding a scaling of the 3 spatial coordinates according to the vector v (negative values codify spatial reflections).

Each component of v must be different from zero.

See also: Transformation

Raytracing.translation — Function

translation(v::Vec) :: Transformation
translation(x::Float64, y::Float64, z::Float64) = translation(Vec(x,y,z))

Encoding a rigid translation of the 3 spatial coordinates according to the vector v, which specifies the amount of shift to be applied along the three axes.

See also: Transformation

Raytracing.inverse — Function

inverse(T::Transformation) :: Transformation

Return the inverse affine transformation of T. This method is very cheap to call.

See also: Transformation

Raytracing.is_consistent — Function

is_consistent(T::Transformation) :: Bool

Check the internal consistency of the input transformation, returning a bool variable indicating whether T.M==T.invM. This method is useful when writing tests.

See also: Transformation