BRDFs and Pigments

Raytracing.UniformPigment — Type

UniformPigment <: Pigment(
    color::RGB{Float32} = RGB{Float32}(0.0, 0.0, 0.0)
)

A uniform pigment. This is the most boring pigment: a uniform hue over the whole surface.

See also: Pigment

Raytracing.CheckeredPigment — Type

CheckeredPigment <: Pigment
    color1::RGB{Float32} = RGB{Float32}(1.0, 1.0, 1.0),
    color2::RGB{Float32} = RGB{Float32}(0.0, 0.0, 0.0),
    num_steps::Int64 = 2
)

A checkered pigment. The number of rows/columns in the checkered pattern is tunable through the integer value num_steps, but you cannot have a different number of repetitions along the u/v directions.

See also: Pigment

Raytracing.ImagePigment — Type

ImagePigment <: Pigment(
    image::HDRimage = HDRimage(3, 2, fill( BLACK, (6,) ) )
)

A textured pigment. The texture is given through a PFM image.

See also: Pigment, HDRimage

Raytracing.DiffuseBRDF — Type

DiffuseBRDF <: BRDF(
    pigment::Pigment = UniformPigment(RGB{Float32}(1.0, 1.0, 1.0)),
    reflectance::Float64 = 1.0
)

A class representing an ideal diffuse BRDF (also called «Lambertian»).

See also: Pigment, UniformPigment

Raytracing.SpecularBRDF — Type

DiffuseBRDF <: BRDF(
    pigment::Pigment = UniformPigment(RGB{Float32}(1.0, 1.0, 1.0)),
    theresold_angle_rad::Float64 = π/180.
)

A class representing an ideal mirror BRDF.

See also: Pigment, UniformPigment

Raytracing.evaluate — Function

evaluate(b::BRDF, n::Normal, in::Vec, out::Vec, uv::Vec2d) :: RGB{Float32}
evaluate(b::DiffuseBRDF, n::Normal, in::Vec, out::Vec, uv::Vec2d) :: RGB{Float32}
evaluate(b::SpecularBRDF, n::Normal, in::Vec, out::Vec, uv::Vec2d) :: RGB{Float32}

Return the RGB color with the specified BRDF b and spatial configuation of durface normal n, incident ray direction in, leaving ray direction out, (u,v) coordinates on surface.

See also: BRDF, DiffuseBRDF, SpecularBRDF Normal, Vec, Vec2d