Demo

Raytracing.demo — Function

demo(
     	renderer::Renderer = FlatRenderer(),
	camera_type::String = "per",
	camera_position::Union{Point, Vec} = Point(-1.,0.,0.), 
 	α::Float64 = 0., 
 	width::Int64 = 640, 
 	height::Int64 = 480, 
 	pfm_output::String = "demo.pfm", 
    	png_output::String = "demo.png",
	samples_per_pixel::Int64 = 0, 
	world_type::String = "A",
	bool_print::Bool = true,
	bool_savepfm::Bool = true,
	ONLY_FOR_TESTS::Bool = false,
      )

demo(x::(Pair{T1,T2} where {T1,T2})...) = 
	demo( parse_demo_settings(  Dict( pair for pair in [x...]) )... )

Creates a demo image with the specified options.

There are two possible demo image "world" to be rendered, specified through the input string type.

The type=="A" demo image world consist in a set of 10 spheres of equal radius 0.1: 8 brown spheres are placed at the verteces of a cube of side 1.0, one green-purple checked sphere is in the center of the lower cube face and another multi-colored sphere is in the center of the left cube face.

The type=="B" demo image world consists in a checked x-y plane, a blue opaque sphere, a red reflecting sphere, and a green oblique reflecting plane, all inside a giant emetting sphere.

The creation of the demo image has the objective to check the correct behaviour of the rendering software, specifically the orientation upside-down and left-right.

Arguments

The following input arguments refers to the first method presented in the signature; it's obviously very uncomfortable to use that method, so it's recommended to take advantage of the second one, which allows to write the input values in a dictionary like syntax with arbitrary order and comfort. See the documentation of parse_demo_settings to learn how to use the keys:

renderer::Renderer = FlatRenderer() : renderer to be used in the rendering, with all the settings already setted (exception made for the world, that will be overridden and created here)

camera_type::String = "per" : set the perspective projection view:

camera_type=="per" -> set PerspectiveCamera (default value)

camera_type=="ort" -> set OrthogonalCamera

camera_position::Union{Point, Vec} = Point(-1.,0.,0.) : set the point of observation in (X,Y,Z) coordinates, as aPointor aVec` input object.

α::Float64 = 0. : angle of rotation IN RADIANTS, relative to the vertical (i.e. z) axis with a right-handed rule convention (clockwise rotation for entering (x,y,z)-axis corresponds to a positive input rotation angle)

width::Int64 = 640 and height::Int64 = 480 : pixel dimensions of the demo image; they must be both even positive integers.

pfm_output::String = "demo.pfm" : name of the output pfm file

png_output::String = "demo.png" : name of the output LDR file

samples_per_pixel::Int64 = 0 : number of rays per pixel to be used (antialiasing); it must be a perfect square positive integer (0, 1, 4, 9, ...) and if is set to 0 (default value) is choosen, no anti-aliasing occurs, and only one pixel-centered ray is fired for each pixel.

world_type::String = "A" : specifies the type of world to be rendered ("A" or "B")

bool_print::Bool = true : specifies if the WIP messages of the demo function should be printed or not (useful option for demo_animation)

bool_savepfm::Bool = true : bool that specifies if the pfm file should be saved or not (useful option for demo_animation)

ONLY_FOR_TESTS::Bool = false : it's a bool variable conceived only to test the correct behaviour of the renderer for the input arguments; if set to true, no rendering is made!

See also: Point, Vec, Renderer OnOffRenderer, FlatRenderer, PathTracer, PointLightRenderer, demo_animation, parse_demo_settings

Raytracing.first_world — Function

first_world() :: World

Render the first world (identified with the string "A").

This world consists in a set of 10 spheres of equal radius 0.1: 8 brown spheres are placed at the verteces of a cube of side 1.0, one green-purple checked sphere is in the center of the lower cube face and another multi-colored sphere is in the center of the left cube face.

See also: World, demo, demo_animation

Raytracing.second_world — Function

second_world() :: World

Render the second world (identified with the string "B").

This world consists in a checked x-y plane, a blue opaque sphere, a red reflecting sphere, and a green oblique reflecting plane, all inside a giant emetting sphere.

See also: World, demo, demo_animation

Raytracing.select_world — Function

select_world(type_world::String) ::Function

Select which demo world is used

Raytracing.parse_demo_settings — Function

parse_demo_settings(dict::Dict{String, T}) where {T}
    :: (
        Renderer, Point, Vec, Float64, Int64, Int64, String, String,
        Int64, String, Bool, Bool, Bool,
        )

Parse a Dict{String, T} where {T} for the demo function.

Input

A dict::Dict{String, T} where {T}

Returns

A tuple (renderer, camera_type, camera_position, α, width, height, pfm, png, samples_per_pixel, world_type, bool_print, bool_savepfm) containing the following variables (the corresponding keys are also showed):

renderer::Renderer = haskey(dict, "renderer") ? dict["renderer"] : dict["%COMMAND%"] : it's the renderer to be used (with a default empy World that will be populated in the demo function); the possible keys are two, and must be used differently:

the dict["renderer"] must contain the renderer itself (i.e. a Renderer object); if this key exists, it has the priority on the latter key

the dict["%COMMAND%"] must contain a string that identifies the type of renderer to be used, i.e. one of the following strings:

"dict["%COMMAND%"]=>onoff" -> OnOffRenderer algorithm

"dict["%COMMAND%"]=>"flat" -> FlatRenderer algorithm (default value)

"dict["%COMMAND%"]=>"pathtracing" -> PathTracer algorithm

"dict["%COMMAND%"]=>"pointlight" -> PointLightRenderer algorithm

Moreover, in this second case, you can specify the options for the correspoding renderer through another dictionary associated with the kkey of the renderer name; that options will be parsed thanks to the corresponding functions. We shows in the next lines the key-value syntax described:

"onoff"=>Dict{String, T} where {T} : parsed with parse_onoff_settings

"flat"=>Dict{String, T} where {T} : parsed with parse_flat_settings

"pathtracer"=>Dict{String, T} where {T} : parsed with parse_pathtracer_settings

"pointlight"=>Dict{String, T} where {T} : parsed with parse_pointlight_settings

camera_type::String = dict["camera_type"] : set the perspective projection view; it must be one of the following values, and this is checked with the string2stringoneof function:

"per" -> set PerspectiveCamera (default value)

"ort" -> set OrthogonalCamera

camera_position::String = typeof(dict["camera_position"]) ∈ [Vec, Point] ? dict["camera_position"] : string2vector(dict["camera_position"]) : "[X, Y, Z]" coordinates of the choosen observation point of view; it can be specified in two ways:

if typeof(dict["camera_position"]) is a Vec or a Point, it's passed as-is

else, it must be a String written in the form "[X, Y, Z]" , and it's parsed through string2vectorto aVec` object

α::String = dict["alpha"] : choosen angle of rotation IN RADIANTS respect to vertical (i.e. z) axis with a right-handed rule convention (clockwise rotation for entering (x,y,z)-axis corresponds to a positive input rotation angle)

width::Int64 = string2evenint64(dict["width"]) : number of pixels on the horizontal axis to be rendered; it's converted through string2evenint64 to a even positive integer.

height::Int64 = string2evenint64(dict["height"]) : number of pixels on the vertical axis to be rendered; it's converted through string2evenint64 to a even positive integer.

pfm::String = dict["set_pfm_name"] : output pfm filename (default "demo.pfm")

png::String = dict["setpngname"]: output LDR filename (default"demo.png"`)

samples_per_pixel::Int64 = dict["samples_per_pixel"] : number of ray to be generated for each pixel, implementing the anti-aliasing algorithm; it must be a perfect integer square (0,1,4,9,...) and this is checked with the string2rootint64 function; if 0 (default value) is choosen, no anti-aliasing occurs, and only one pixel-centered ray is fired for each pixel.

world_type::String = dict["world_type"] : type of the world to be rendered; it must be "A" or "B", and this is checked with the string2stringoneof function

bool_print::Bool = dict["bool_print"] : if true (default value), WIP message of demo function are printed (otherwise no; it's useful for the demo_animation function)

bool_savepfm::Bool = dict["bool_savepfm"] : if true (default value), demo function saves the pfm file to disk (otherwise no; it's useful for the demo_animation function)

ONLY_FOR_TESTS::Bool = dict["ONLY_FOR_TESTS"] : it's a bool variable conceived only to test the correct behaviour of the renderer for the input arguments; if set to true, no rendering is made!

See also: demo, demo_animation, Renderer, Vec, string2evenint64, string2stringoneof, string2positive, string2vector, string2rootint64

Demo Animation

Raytracing.demo_animation — Function

demo_animation( 
		renderer::Renderer = FlatRenderer(),
		camera_type::String = "per",
    		width::Int64 = 200, 
    		height::Int64 = 150,
   		anim_output::String = "demo-animation.mp4",
		samples_per_pixel::Int64 = 0,
		world_type::String = "A", 
		ONLY_FOR_TESTS::Bool = false,
	)

demo_animation(x::(Pair{T1,T2} where {T1,T2})...) = 
	demo_animation( parse_demoanimation_settings(  Dict( pair for pair in [x...]) )... )

Creates an animation of the demo image with the specified options. It's necessary to have istalled the ffmpeg software to run this function.

This function works following this steps:

creates an hidden directory, called ".wip_animation"; if it already exists, it will be destroyed and recreated.

inside ".wpi_animation", creates 360 png images of the demo image (using the demo function with the specified projection, renderer and image dims); each image correspons to a frame of the future animation

through the ffmpeg software, the 360 png images are converted into the animation mp4 file, and saved in the main directory

the ".wpi_animation" directory and all the png images inside it are destroyed

Arguments

The following input arguments refers to the first method presented in the signature; it's obviously very uncomfortable to use that method, so it's recommended to take advantage of the second one, which allows to write the input values in a dictionary like syntax with arbitrary order and comfort. See the documentation of parse_demoanimation_settings to learn how to use the keys:

renderer::Renderer = FlatRenderer() : renderer to be used in the rendering, with all the settings already setted (exception made for the world, that will be overridden and created here)

camera_type::String = "per" : set the perspective projection view:

camera_type=="per" -> set PerspectiveCamera (default value)

camera_type=="ort" -> set OrthogonalCamera

camera_position::Union{Point, Vec} = Point(-1.,0.,0.) : set the point of observation in (X,Y,Z) coordinates, as aPointor aVec` input object.

width::Int64 = 640 and height::Int64 = 480 : pixel dimensions of the demo image; they must be both even positive integers.

pfm_output::String = "demo.pfm" : name of the output pfm file

png_output::String = "demo.png" : name of the output LDR file

samples_per_pixel::Int64 = 0 : number of rays per pixel to be used (antialiasing); it must be a perfect square positive integer (0, 1, 4, 9, ...) and if is set to 0 (default value) is choosen, no anti-aliasing occurs, and only one pixel-centered ray is fired for each pixel.

world_type::String = "A" : specifies the type of world to be rendered ("A" or "B")

ONLY_FOR_TESTS::Bool = false : it's a bool variable conceived only to test the correct behaviour of the renderer for the input arguments; if set to true, no rendering is made!

See also: Point, Vec, Renderer OnOffRenderer, FlatRenderer, PathTracer, PointLightRenderer, demo, parse_demoanimation_settings

Raytracing.parse_demoanimation_settings — Function

parse_demoanimation_settings(dict::Dict{String, T}) where {T}
    :: (
        Renderer, String, Vec, Int64, Int64, String,
        Int64, String, Bool,
        )

Parse a Dict{String, T} where {T} for the demo_animation function.

Input

A dict::Dict{String, T} where {T}

Returns

A tuple (renderer, camera_type, camera_position, width, height, anim, samples_per_pixel, world_type) containing the following variables (the corresponding keys are also showed):

renderer::Renderer = haskey(dict, "renderer") ? dict["renderer"] : dict["%COMMAND%"] : it's the renderer to be used (with a default empy World that will be populated in the demo function); the possible keys are two, and must be used differently:

the dict["renderer"] must contain the renderer itself (i.e. a Renderer object); if this key exists, it has the priority on the latter key

the dict["%COMMAND%"] must contain a string that identifies the type of renderer to be used, i.e. one of the following strings:

"dict["%COMMAND%"]=>onoff" -> OnOffRenderer algorithm

"dict["%COMMAND%"]=>"flat" -> FlatRenderer algorithm (default value)

"dict["%COMMAND%"]=>"pathtracing" -> PathTracer algorithm

"dict["%COMMAND%"]=>"pointlight" -> PointLightRenderer algorithm

Moreover, in this second case, you can specify the options for the correspoding renderer through another dictionary associated with the kkey of the renderer name; that options will be parsed thanks to the corresponding functions. We shows in the next lines the key-value syntax described:

"onoff"=>Dict{String, T} where {T} : parsed with parse_onoff_settings

"flat"=>Dict{String, T} where {T} : parsed with parse_flat_settings

"pathtracer"=>Dict{String, T} where {T} : parsed with parse_pathtracer_settings

"pointlight"=>Dict{String, T} where {T} : parsed with parse_pointlight_settings

camera_type::String = dict["camera_type"] : set the perspective projection view; it must be one of the following values, and this is checked with the string2stringoneof function:

"per" -> set PerspectiveCamera (default value)

"ort" -> set OrthogonalCamera

camera_position::String = typeof(dict["camera_position"]) ∈ [Vec, Point] ? dict["camera_position"] : string2vector(dict["camera_position"]) : "[X, Y, Z]" coordinates of the choosen observation point of view; it can be specified in two ways:

if typeof(dict["camera_position"]) is a Vec or a Point, it's passed as-is

else, it must be a String written in the form "[X, Y, Z]" , and it's parsed through string2vectorto aVec` object

width::Int64 = string2evenint64(dict["width"]) : number of pixels on the horizontal axis to be rendered; it's converted through string2evenint64 to a even positive integer.

height::Int64 = string2evenint64(dict["height"]) : number of pixels on the vertical axis to be rendered; it's converted through string2evenint64 to a even positive integer.

anim::String = dict["set_anim_name"] : output animation filename (default "demo_animation.mp4")

samples_per_pixel::Int64 = dict["samples_per_pixel"] : number of ray to be generated for each pixel, implementing the anti-aliasing algorithm; it must be a perfect integer square (0,1,4,9,...) and this is checked with the string2rootint64 function; if 0 (default value) is choosen, no anti-aliasing occurs, and only one pixel-centered ray is fired for each pixel.

world_type::String = dict["world_type"] : type of the world to be rendered; it must be "A" or "B", and this is checked with the string2stringoneof function

ONLY_FOR_TESTS::Bool = dict["ONLY_FOR_TESTS"] : it's a bool variable conceived only to test the correct behaviour of the renderer for the input arguments; if set to true, no rendering is made!

See also: demo, demo_animation, Renderer, Vec, string2evenint64, string2stringoneof, string2positive, string2vector, string2rootint64