Avaiable shapes

Raytracing.ray_intersection — Method

ray_intersection(shape::Shape, ray::Ray) :: Union{HitRecord, Nothing}

Compute the intersection between a Ray and a Shape.

See also: Ray, Shape

Raytracing.ray_intersection — Method

ray_intersection(world::World, ray::Ray) :: Union{HitRecord, Nothing}

Determine whether the ray intersects any of the objects of the given world. Return a HitRecord, or nothing if no intersection is found.

See also: Ray, World, HitRecord

Sphere

Raytracing.Sphere — Type

Sphere <: Shape(
    T::Transformation = Transformation(),
    Material::Material = Material()
)

A 3D unit sphere, i.e. centered on the origin of the axes and with radius 1.0.

Arguments

T::Transformation : transformation associated to the sphere.

Material::Material : material that constitutes the sphere.

See also: Shape, Transformation, Material

Raytracing.sphere_point_to_uv — Function

sphere_point_to_uv(point::Point) :: Vec2d

Convert a 3D point $P = (P_x, P_y, P_z)$ on the surface of the unit sphere into a 2D Vec2d using the following spherical coordinates:

\[u = \frac{\phi}{2\pi} = \frac{\arctan (P_y / P_x)}{2\pi}, \quad v = \frac{\theta}{\pi} = \frac{\arccos (P_z)}{\pi}\]

See also: Point, Vec2d, Sphere

Raytracing.sphere_normal — Function

sphere_normal(point::Point, ray_dir::Vec) :: Normal

Compute the Normal of a unit sphere.

The normal is computed for the given Point $point = (P_x, P_y, P_z)$ (with $\sqrt{P_x^2 + P_y^2 + P_z^2}=1$) on the surface of the sphere, and it is chosen so that it is always in the opposite direction with respect to the given Vec ray_dir.

See also: Point, Ray, Normal, Sphere

Raytracing.ray_intersection — Method

ray_intersection(sphere::Sphere, ray::Ray) :: Union{HitRecord, Nothing}

Check if the ray intersects the sphere. Return a HitRecord, or nothing if no intersection is found.

See also: Ray, Sphere, HitRecord

Plane

Raytracing.Plane — Type

Plane <: Shape(
    T::Transformation = Transformation(),
    Material::Material = Material()
)

A 3D unit plane, i.e. the x-y plane (set of 3D points with z=0).

Arguments

T::Transformation : transformation associated to the plane.

Material::Material : material that constitutes the plane.

See also: Shape, Transformation, Material

Raytracing.plane_point_to_uv — Function

plane_point_to_uv(point::Point) :: Vec2d

Convert a 3D point $P = (P_x, P_y, P_z)$ on the surface of the unit plane into a 2D Vec2d using the following periodical coordinates:

\[u = P_x - \lfloor P_x \rfloor, \quad v = P_y - \lfloor P_y \rfloor,\]

where $\lfloor \cdot \rfloor$ indicates the rounding down approximation, in order to guarantee that $u, v \in [0, 1)$.

See also: Point, Vec2d, Plane

Raytracing.plane_normal — Function

plane_normal(point::Point, ray_dir::Vec) :: Normal

Compute the Normal of a unit plane.

The normal is computed for the given point $P = (P_x, P_y, 0)$ on the surface of the plane, and it is chosen so that it is always in the opposite direction with respect to the given Vec ray_dir.

See also: Point, Ray, Normal, Plane

Raytracing.ray_intersection — Method

ray_intersection(plane::Plane, ray::Ray) :: Union{HitRecord, Nothing}

Check if the ray intersects the plane. Return a HitRecord, or nothing if no intersection is found.

See also: Ray, Plane, HitRecord

Torus

Raytracing.Torus — Type

Torus <: Shape(
    T::Transformation = Transformation()
    Material::Material = Material()
    r::Float64 = 0.5
    R::Float64 = 1.0
)

A 3D unit torus, a ring with circular section; has origin in (0, 0, 0) and axis parallel to the y-axis.

Arguments

T::Transformation: transformation associated to the torus.

Material::Material : material that constitutes the torus.

r::Float64 : radius of the circular section.

R::Float64 : distance between the torus center and the section center.

^ ̂y                __-__
|                 /     \ 
|---O------------(---o---)
|                 \__ __/
|                    -
      <--------R------><-r->

See also: Shape, Transformation, Material

Missing docstring.

Missing docstring for Raytracing.torus_point_to_uv. Check Documenter's build log for details.

Raytracing.torus_normal — Function

torus_normal(p::Point, ray_dir::Vec, R::Float64) -> Normal

Compite the Normal of a torus

The normal is computed for Point (a point on the surface of the torus), and it is chosen so that it is always in the opposite direction with respect to ray_dir (Vec).

Raytracing.ray_intersection — Method

ray_intersection(shape::Shape, ray::Ray) :: Union{HitRecord, Nothing}

Compute the intersection between a Ray and a Shape.

See also: Ray, Shape

Cube

Raytracing.Cube — Type

Cube <: Shape(
    T::Transformation = Transformation(),
    Material::Material = Material()
)

A 3D unit cube, i.e. an axis aligned cube with side 1 centered in the origin.

Arguments

T::Transformation : transformation associated to the cube.

Material::Material : material that constitutes the cube.

See also: Shape, Transformation, Material

Raytracing.cube_point_to_uv — Function

cube_point_to_uv(point::Point) :: Vec2d

Convert a 3D point $P = (P_x, P_y, P_z)$ on the surface of the unit cube into a 2D Vec2d using the following coordinates:

\[P_x = \frac{1}{2} \lor P_x = -\frac{1}{2} \quad \Rightarrow \quad u = P_y + \frac{1}{2} \; , \; v = P_z + \frac{1}{2}\]

\[P_y = \frac{1}{2} \lor P_y = -\frac{1}{2} \quad \Rightarrow \quad u = P_x + \frac{1}{2} \; , \; v = P_z + \frac{1}{2}\]

\[P_z = \frac{1}{2} \lor P_z = -\frac{1}{2} \quad \Rightarrow \quad u = P_x + \frac{1}{2} \; , \; v = P_y + \frac{1}{2}\]

\[P_x \neq \frac{1}{2}, -\frac{1}{2} \land P_y \neq \frac{1}{2}, -\frac{1}{2} \land P_z \neq \frac{1}{2}, -\frac{1}{2} \quad \Rightarrow \quad \mathrm{throw Exception}\]

See also: Point, Vec2d, Cube

Raytracing.cube_normal — Function

cube_normal(point::Point, ray_dir::Vec) :: Normal

Compute the Normal of a unit cube.

The normal is computed for the given point on the surface of the cube, and it is chosen so that it is always in the opposite direction with respect to the given Vec ray_dir.

See also: Point, Ray, Normal, Cube

Raytracing.ray_intersection — Method

ray_intersection(cube::Cube, ray::Ray) :: Union{HitRecord, Nothing}

Check if the ray intersects the cube. Return a HitRecord, or nothing if no intersection is found.

The implementation is only a long boring list of if-else block statements, and may have to be optimized.

See also: Ray, Cube, HitRecord

Axis-Aligned Bounding Box

Raytracing.AABB — Type

AABB <: Shape(
    vertexes::SVector{6, Float32}
)

An Axis-Aligned Boundary Box.

This shape is not conceived to be rendered in the image, it has the only purpose to optimize the various ray_intersection functions.

Arguments

vertexes::SVector{6, Float32} : the 6 coordinates of the exterior vertexes defining the AABB

See also: Shape, ray_intersection

Raytracing.ray_intersection — Method

ray_intersection(AABB::AABB, ray::Ray) :: Bool

Check if the ray intersects the AABB. Return true if intersection occurs, false otherwise.

See also: Ray, AABB, HitRecord

Triangle

Raytracing.Triangle — Type

Triangle <: Shape(
    vertexes::SVector{3, Point} = 
        SVector{3, Point}(Point(√3/2, 0, 0), Point(0, 0.5, 0), Point(0, -0.5, 0)),
    Material::Material = Material()
)

A 3D triangle.

Arguments

vertexes::SVector{3, Point} : points associated to the triangle.

Material::Material : material that constitutes the triangle.

See also: Shape, Material

Raytracing.triangle_point_to_uv — Function

triangle_point_to_uv(triangle::Triangle, point::Point) :: Vec2d

Return the barycentic coordinates of the given point for the input triangle.

If the triangle is made of the vertexes $(A,B,C)$ (memorized in this order), then the point $P$ has coordinates $(u,v) = (\beta, \gamma)$ such that:

\[ P(\beta, \gamma) = A + \beta \,(B - A) + \gamma \,(C-A)\]

The analitic resolution of this linear system is:

\[\begin{aligned} &\beta = \frac{ (P_x - A_x)(C_y - A_y) - (P_y - A_y)(C_x - A_x) }{ (B_x - A_x)(C_y - A_y) - (B_y - A_y)(C_x - A_x) } \\ &\gamma = \frac{ (P_x - A_x)(B_y - A_y) - (P_y - A_y)(B_x - A_x) }{ (C_x - A_x)(B_y - A_y) - (C_y - A_y)(B_x - A_x) } \end{aligned}\]

NOTE: this function do not check if $P$ is on the plane defined by $(A,B,C)$, neither if $P$ is inside the triangle made of them!

See also: Triangle, Vec2d, Point

Raytracing.triangle_barycenter — Function

triangle_barycenter(triangle::Triangle) :: Point

Return the barycenter of the given triangle.

For a triangle with vertexes $(A, B, C)$, the barycenter is $M$:

\[ M = \frac{A + B + C}{3}\]

See also: Triangle, Point

Raytracing.triangle_normal — Function

triangle_normal(triangle::Triangle, ray_dir::Vec) :: Normal

Compute the Normal of a given triangle.

The normal for a triangle with vertexes $(A, B, C)$ is computed as follows:

\[ n = \pm (B-A) \times (C-A)\]

where the sign is chosen so that it is always in the opposite direction with respect to the given ray_dir.

See also: Point, Ray, Normal, Triangle

Raytracing.ray_intersection — Method

ray_intersection(triangle::Triangle, ray::Ray) :: Union{HitRecord, Nothing}

Check if the ray intersects the triangle. Return a HitRecord, or nothing if no intersection is found.

For a triangle with vertexes $(A, B, C)$ and a ray defined with the simple equation $r(t) = O + t \, \vec{d}$, the coordinates $(u,v) = (\beta, \gamma)$ and the t value of intersection are obtained solving this linear system:

\[ \begin{bmatrix} B_x-A_x & C_x-A_x & -d_x \\ B_y-A_y & C_y-A_y & -d_y \\ B_z-A_z & C_z-A_z & -d_z \end{bmatrix} \begin{bmatrix} u \\ v \\ t \end{bmatrix} = \begin{bmatrix} O_x - A_x\\ O_z - A_z\\ O_z - A_z \end{bmatrix}\]

See also: Ray, Triangle, HitRecord