Avaiable shapes

Sphere

Raytracing.Sphere — Type

Sphere <: Shape(
    T::Transformation = Transformation(),
    Material::Material = Material()
)

A 3D unit sphere, i.e. centered on the origin of the axes and with radius 1.0.

Arguments

T::Transformation : transformation associated to the sphere.

Material::Material : material that constitutes the sphere.

See also: Shape, Transformation, Material

Raytracing.sphere_point_to_uv — Function

sphere_point_to_uv(point::Point) :: Vec2d

Convert a 3D point $P = (P_x, P_y, P_z)$ on the surface of the unit sphere into a 2D Vec2d using the following spherical coordinates:

\[u = \frac{\phi}{2\pi} = \frac{\arctan (P_y / P_x)}{2\pi}, \quad v = \frac{\theta}{\pi} = \frac{\arccos (P_z)}{\pi}\]

See also: Point, Vec2d, Sphere

Raytracing.sphere_normal — Function

sphere_normal(point::Point, ray_dir::Vec) :: Normal

Compute the Normal of a unit sphere.

The normal is computed for the given Point $point = (P_x, P_y, P_z)$ (with $\sqrt{P_x^2 + P_y^2 + P_z^2}=1$) on the surface of the sphere, and it is chosen so that it is always in the opposite direction with respect to the given Vec ray_dir.

See also: Point, Ray, Normal, Sphere

Raytracing.ray_intersection — Method

ray_intersection(sphere::Sphere, ray::Ray) :: Union{HitRecord, Nothing}

Check if the ray intersects the sphere. Return a HitRecord, or nothing if no intersection is found.

See also: Ray, Sphere, HitRecord

Plane

Raytracing.Plane — Type

Plane <: Shape(
    T::Transformation = Transformation(),
    Material::Material = Material()
)

A 3D unit plane, i.e. the x-y plane (set of 3D points with z=0).

Arguments

T::Transformation : transformation associated to the plane.

Material::Material : material that constitutes the plane.

See also: Shape, Transformation, Material

Raytracing.plane_point_to_uv — Function

plane_point_to_uv(point::Point) :: Vec2d

Convert a 3D point $P = (P_x, P_y, P_z)$ on the surface of the unit plane into a 2D Vec2d using the following periodical coordinates:

\[u = P_x - \lfloor P_x \rfloor, \quad v = P_y - \lfloor P_y \rfloor,\]

where $\lfloor \cdot \rfloor$ indicates the rounding down approximation, in order to guarantee that $u, v \in [0, 1)$.

See also: Point, Vec2d, Plane

Raytracing.plane_normal — Function

plane_normal(point::Point, ray_dir::Vec) :: Normal

Compute the Normal of a unit plane.

The normal is computed for the given Point $point = (P_x, P_y, 0)$ on the surface of the plane, and it is chosen so that it is always in the opposite direction with respect to the given Vec ray_dir.

See also: Point, Ray, Normal, Plane

Raytracing.ray_intersection — Method

ray_intersection(plane::Plane, ray::Ray) :: Union{HitRecord, Nothing}

Check if the ray intersects the plane. Return a HitRecord, or nothing if no intersection is found.

See also: Ray, Plane, HitRecord